% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{as_ode}
\alias{as_ode}
\title{Interpret a model as an ODE}
\usage{
as_ode(m, cla = requireNamespace("pracma", quietly = TRUE))
}
\arguments{
\item{m}{a list of data.frames, each corresponding to a TSV file or
sheet in a spreadsheet.}

\item{cla}{a Boolean value indicating whether conservation law
analysis should be performed.}
}
\value{
a list that contains a summary of this model interpreted as
an ODE, crucially, the list contains the element \code{vf}, the
right-hand-side (vector field) of the ODE, this is the main
result of this function.
}
\description{
This function accepts a list generated from a collection of TSV
files (or a similar format) and interprets the contents as an
ordinary differential equation (ODE).
}
\details{
The argument \code{m} can be obtained via \code{model_from_tsv()}.
It has the components:
\itemize{
\item \code{m$Constant}
\item \code{m$Parameter}
\item \code{m$Input}
\item \code{m$Expression}
\item \code{m$Compound}
\item \code{m$Reaction}
\item \code{m$Experiment}
}

There can be additional components describing measured data for
this model.
}
\examples{
f <- uqsa_example("AKAR4")
m <- model_from_tsv(f)
o <- as_ode(m)
print(names(o))
print(o$vf)
}
