% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{base<-}
\alias{base<-}
\title{Convert to linear space}
\usage{
base(x, i = seq_along(x)) <- value
}
\arguments{
\item{x}{a numeric vector}

\item{i}{a subset of values in x, defaults to all values of x}

\item{value}{the base of the logarithm x was provided in}
}
\value{
x will be changed to be in linear space
}
\description{
A number given in some logarithmic space can be transformed back to linear space
A call like \code{base(x) <- 10} means that x was provided in base-10 logarithm form.
This will adjust \code{x} so that it is now in linear space.
}
\details{
If \code{x} was provided in logarithmic space, then it is an exponent to
the given base (value).
}
\examples{
 x <- 2
 base(x) <- 10
 print(x)
}
