% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{c.path<-}
\alias{c.path<-}
\title{Add information about the model's C code}
\usage{
\method{c}{`path<-`}(o, value)
}
\arguments{
\item{o}{the ode (list of named arrays and matrices)}

\item{value}{the path to the compiled model}
}
\value{
modified o, with information about compiled code
}
\description{
Adds the location of the model's C code (a file).
}
