% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{c.path}
\alias{c.path}
\title{Retrieve information about the model's C code}
\usage{
\method{c}{path}(o)
}
\arguments{
\item{o}{the ODE, or CME model}
}
\value{
the path where the c code is stored
}
\description{
Returns the location of the model's C code (a file).
}
