% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{checkFitWithPreviousExperiments}
\alias{checkFitWithPreviousExperiments}
\title{ABC acceptance of currently sampled values given old data (Prior)}
\usage{
checkFitWithPreviousExperiments(draws, objectiveFunction, delta)
}
\arguments{
\item{draws}{matrix of sampled values (to be filtered).}

\item{objectiveFunction}{function that, given a (vectorial)
parameter as input, simulates the model, and outputs the
distance between experimental data and data simulated from the
model with the parameter provided in input}

\item{delta}{the acceptance threshold.}
}
\value{
a filtered subset of acceptable parameter draws
}
\description{
The prior probability density model using copulas and vines is not
perfect, so values sampled from an imperfect prior estimate can be
checked against old data.
}
