% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{clear_yacas_environment}
\alias{clear_yacas_environment}
\title{Clear Yacas variables}
\usage{
clear_yacas_environment()
}
\value{
list of variables cleared as a character array
}
\description{
The \code{reset} operation doesn't work in yacas, so this function wipes
every variable one by one.
}
\examples{
Ryacas::yac_str("y := 2*x")
Ryacas::yac_str("restart")
print(Ryacas::yac_str("D(x) y^2"))
clear_yacas_environment()
print(Ryacas::yac_str("D(x) y^2"))
}
