% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{cMath}
\alias{cMath}
\title{Make math C compatible}
\usage{
cMath(v, gsl = TRUE)
}
\arguments{
\item{v}{a character vector with math expressions}

\item{gsl}{use power functions from the GNU scientific library}
}
\value{
a vector with c-compatible math
}
\description{
Given a character array with math expressions, return a similar
vector, with c-compatible math. The most important change is
related to powers.
}
