% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{column}
\alias{column}
\title{Get j-th column with names}
\usage{
column(m, j = 1)
}
\arguments{
\item{m}{a matrix}

\item{j}{a column index}
}
\value{
a named vector
}
\description{
When indexing a matrix, rownames are lost. This function will
return a column of a matrix, as a vector (dropping rank so to
speak), but the vector will retain the rownames of the matrix
}
