% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{convert.parameter}
\alias{convert.parameter}
\title{convert ODE parameter to Gillespie parameter}
\usage{
convert.parameter(k, n = NULL, LV = 602214076)
}
\arguments{
\item{k}{the ODE reaction rate coefficient (mandatory)}

\item{n}{multiplicity of each reactant, if any (order > 0)}

\item{LV}{L*V -- product of _Avogadro's number_ and _volume_ [defaults to 6.02214076e+8]}
}
\value{
rescaled parameter for stochastic simulation with a comment of how to re-scale it
}
\description{
ODE parameters usually have a different unit of measurement than
the parameters we need for stochastic simulators.  ODEs have
fluxes, which are in multiples of M/s (M is mol/liter), same unit
as the first derivative of the state variables.
}
\details{
The reaction rate coefficients of mass action kinetics, kf and kb
have units that are compatible with the flux units, depending on
the order of the reaction (the order is related to the reaction's
stoichiometry).
}
\examples{
reaction: "2 A + B -> C"
k <- 1.0
attr(k,'unit') <- "µM/s"
n <- c(2,1)
reactants <- c('A','B')
convert.parameter(k,n)
}
