% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{defaultParMap}
\alias{defaultParMap}
\title{Default parameter mapping used by the MCMC module}
\usage{
defaultParMap(parMCMC)
}
\arguments{
\item{parMCMC}{the sampling variables (numeric vector)}
}
\description{
This map is used by the simulator to transform sampling variables
into ODE-model porameters.
}
