% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{defaultParMapJac}
\alias{defaultParMapJac}
\title{Default parameter mapping, jacobian}
\usage{
defaultParMapJac(parMCMC)
}
\arguments{
\item{parMCMC}{the sampling variables (numeric vector)}
}
\description{
This map is used by the simulator to transform sampling variables
into ODE-model porameters. As we often calculate sensitivites, we
alos need the jacobian of the map, due to the chain rule of
differentiation.
}
