% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{determinePrefix}
\alias{determinePrefix}
\title{Determine a prefix from a character vector str of similar contents}
\usage{
determinePrefix(str, split = "-", collapse = "-")
}
\arguments{
\item{str}{a character vector}

\item{split}{the token to use for strsplit instead of '-', this should be \code{character(0)} if you want to split letter by letter}
}
\value{
the prefix common to all entries of str.
}
\description{
The result is such that \code{all(startsWith(str,determinePrefix(str)))} is \code{TRUE}.
}
\details{
By default, the strings are assumed to be '-' separated words, and
a series of words is found to be thje prefix if all entries start
with that set of words.

The normal case is c("abc-1","abc-2b","abc-2a") maps to "abc"
}
