% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{.unit.kind}
\alias{.unit.kind}
\title{find unit category}
\usage{
.unit.kind(kind)
}
\arguments{
\item{kind}{the unnormalized string that humans use to write a unit
(but without prefix)}
}
\value{
normalized category name of the unit kind: litre, mole,
    metre, kilogram, gram, ampere, candela, second, kelvin, hour,
    molarity, dimensionless. defaults to dimensionless.
}
\description{
This function reads a unit, without SI prefix, and returns a string
from a smaller subset of unit kinds, similar to what is defined in
SBML. This normalizes the various ways to write the same unit:
"meter", "m" and "metre".
}
\details{
The unit kind of "m" is "metre", the kind of "g" is "gram".
}
\examples{
>  .unit.kind("meter")
[1] "metre"
}
