% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{.unit.scale}
\alias{.unit.scale}
\title{Unit scale from SI prefix}
\usage{
.unit.scale(prefix)
}
\arguments{
\item{prefix}{a string, e.g.: "M", "mega", "m", "milli", "µ", "micro", etc.}
}
\value{
an integer that corresponds to the prefix, defaults to 0.
}
\description{
This function reads a prefix from a string and returns the exponent
(base-10) that this prefix represents.
}
\examples{
> .unit.scale("M")
[1] 6
> .unit.scale("µ")
[1] -6
}
