% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{dUniformPrior}
\alias{dUniformPrior}
\title{dUniformPrior creates a uniform density function}
\usage{
dUniformPrior(ll, ul)
}
\arguments{
\item{ll}{lower limit of the random variables (a vector)}

\item{ul}{upper limit of the random variables (same size vector as ll)}
}
\value{
a probability density function on vectors with the same length as ll and ul.
}
\description{
The returned density function takes vectors of the same size as ll
and ul. It returns the product of the component's one-dimensional
uniform distributions.
}
\examples{
dup<-dUniformPrior(ll=c(0,1,2),ul=c(1,2,3))
dup(c(0.5,1.5,2.5))
}
