% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLD.R
\name{exact_normal_kld}
\alias{exact_normal_kld}
\title{Multivariate Normal Distribution KLD}
\usage{
exact_normal_kld(muA, SigmaA, muB, SigmaB)
}
\arguments{
\item{muA}{mean of distribution A}

\item{SigmaA}{Covariance of distribution A}

\item{muB}{mean of distribution B}

\item{SigmaB}{Covariance of distribution B}
}
\value{
the KLD value D(A|B)
}
\description{
Like the \link{KLD} function, this function calculates KLD values, but
for the specific case of multivariate normal distributions
$\mathcal{N}\emph{{A}$ and $\mathcal{N}}{B}$.  The two distributions are
specified using $\mu$ and $\Sigma$ values (mean and covariance).
}
