% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{find_parameter_reaction}
\alias{find_parameter_reaction}
\title{Find the stoichiometry for a given parameter name}
\usage{
find_parameter_reaction(p, reactants, products, kinetic.law)
}
\arguments{
\item{p}{parameter name}

\item{reactants}{stoichiometry of the reaction's left side}

\item{products}{stoichiometry of the reaction's right side}

\item{kinetic.law}{a character matrix of fluxes, column 1 for
forward reaction, column2 for backward reactions}
}
\value{
the stoichiometry entry that belongs to the given parameter
}
\description{
Given a list representation of stoichiometry (list of named integer
vectors), and a character vector of kinetic laws, this function
returns the correct stoichiometry for the given parameter.
}
