% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{fisherInformation}
\alias{fisherInformation}
\title{Fisher Information from Sensitivity}
\usage{
fisherInformation(model, experiments, parMap = identity, parMapJac = 1)
}
\arguments{
\item{model}{list of R functions for the ODE model}

\item{experiments}{list of experiments, with inputs}

\item{parMap}{mapping between MCMC variables and ODE parameters}

\item{parMapJac}{the jacobian of the above map}
}
\value{
fisher information calculating funciton
}
\description{
Given a list of simulation sensitivities, this function returns the
fisher information (sum over all experiments). The actual work is
done in the returned function that implicitly depends on the model,
experiments, and parameter mapping
}
\details{
return value: function(par, simulations, sensitivity) ->
fisherInformation (matrix)

where par refers to the model parameters
(possibly transformed), and simulations performed with those
parameters.
}
