% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{fisherInformationFromGSA}
\alias{fisherInformationFromGSA}
\title{Calculate Global Fisher Information}
\usage{
fisherInformationFromGSA(Sample, yf = NULL, E)
}
\arguments{
\item{Sample}{an MCMC sample, or ABC sample}

\item{yf}{the simulations of the sample}

\item{E}{experiments list}
}
\description{
Given a sample, this performs global sensitivity analysis, and then
squares the sensitivity.
}
