% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{flux_matrix}
\alias{flux_matrix}
\title{Creates a Matrix with reaction kinetic entries}
\usage{
flux_matrix(Reaction)
}
\arguments{
\item{Reaction}{a data.frame with a column named 'kinetic.law'}
}
\description{
Given a data.frame that describes reaction, this function extracts
the forward flux and backward flux from the kinetic.law column.
}
