% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{formulae}
\alias{formulae}
\title{Find a column that contains some kind of mathematic expression in a data.frame}
\usage{
formulae(df)
}
\arguments{
\item{df}{a data.frame with a "formula" column}
}
\value{
character vector with names taken from the row names of df
}
\description{
Given a data.frame that should contain a column that assigns a math
expression to a name (in row names), this function returns a named
character vector with the expressions. The formula column should be
named "formula" (if it exists, only this column will be used). But
some other spellings will also work as fallback. As a fallback
"value" is acceptable as well, because it makes sense to say
"the value of x is 'y/2+1'", even though it is not an atomic value
(but an expression).
}
\examples{
df <- data.frame(formula=c("exp(x)","10^x","2*x + 3"),row.names=c("f1","f2",'f3'))
formulae(df)
df <- data.frame(value=c("exp(x)","10^x","2*x + 3"),row.names=c("f1","f2",'f3'))
formulae(df)
}
