% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{ftsplit}
\alias{ftsplit}
\title{splits a scalar string into parts}
\usage{
ftsplit(str, s)
}
\arguments{
\item{str}{a string (character vector of length 1)}

\item{s}{a split token}
}
\value{
a character vector of the components without leading or trailing whitespace
}
\description{
This function splits a string like strsplit, but it removes
whitespace from the result (on both ends) and the split token is
taken literally
}
\examples{
ftsplit(" A + 2*B ","+")
returns c("A","2*B")
}
