% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{generate_code}
\alias{generate_code}
\title{Construct Code}
\usage{
generate_code(Model, language = "C", LV = 602214076)
}
\arguments{
\item{Model}{either CME or ODE model}

\item{language}{either C or R}

\item{LV}{Avogadro's constant multiplied by the system's volume in
litres, only used for CME models}
}
\value{
a character vector with the code
}
\description{
Interpret the first argument and generate code in the specified
language for the model type.
}
\details{
Whenever the model \code{Model} is of type \code{"cme"}, the LV parameter is
used to determine the actual number of molecules in the
system. Otherwise it is ignored.
}
