% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{generateGillespieCode}
\alias{generateGillespieCode}
\title{Generate C Code to solve a model stochastically}
\usage{
generateGillespieCode(sb, LV = 602214076)
}
\arguments{
\item{sb}{SBtab list of data-frames, as returned by
SBtabVFGEN::sbtab_from_*() functions}

\item{LV}{Avogadro's Constant * volume \link{in litres}}
}
\value{
character vector with code
}
\description{
This function tries to generate code for a stochastic solver, but
assumming that the SBtab model is written in terms of
concentrations and rate coefficients.
}
\details{
The model is interpreted by \code{stochasticModel()}, parameters found
in reaction kinetics are converted to stochastic parameters.
Input parameters are not converted.
}
