% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{generateRCode}
\alias{generateRCode}
\title{Write R code}
\usage{
generateRCode(odeModel)
}
\arguments{
\item{odeModel}{a list, as returned from SBtabVFGEN::sbtab_to_vfgen()}
}
\value{
a character vector with the generated code, one vector-element is one line of code.
}
\description{
This function expects a list of character vectors, as returned by
SBtabVFGEN::sbtab_to_vfgen. This list describes an ODE model
(initial values, default parameters, transformation events, output
functions).  This function uses this information, calculates
Jacobians via Ryacas and returns a character vector with R source
code for the deSolve package.
}
\details{
The value can be written to a file:
\code{cat(generateRCode(odeModel),sep="\\n",file=...)}. This file can be
sourced later.
}
