% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationsAndData.R
\name{getDose}
\alias{getDose}
\title{Get the values of the input for a series of dose response experiments}
\usage{
getDose(experiments)
}
\arguments{
\item{experiments}{list of experimental data from the same dose response experiment}
}
\value{
vector of inputs (i.e. dose) that varies among the experiments provided to the function.
        The name of the input is saved in "comment(dose)".
}
\description{
This function finds the vector of inputs that  varies among a series of experiments
that are part of a dose response experiment
}
