% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCalibration.R
\name{getMCMCPar}
\alias{getMCMCPar}
\title{Selects MCMC scheme specific setup parameters}
\usage{
getMCMCPar(prePar, preDelta, p = 0.05, sfactor = 0.1, delta = 0.01, num = 1)
}
\arguments{
\item{prePar}{a sample of parameters from pre-Calibration}

\item{preDelta}{distance values (scores) for those parameters}

\item{p}{fraction (top scoring) of sampled points to base Sigma on \link{default 0.05}}

\item{sfactor}{scales Sigma up or down \link{default 0.1}}

\item{delta}{ABC threshold \link{default 1e-2}}

\item{num}{number of different starting parameter vectors \link{default 1}.}
}
\value{
Sigma and startPar (matrix with \code{num} rows) as a list
}
\description{
The MCMC scheme uses a transition kernel. This function returns the
parameters of that transition kernel. Better parameters make the
Markov chain perform better (i.e. lower auto-correlation).
}
