% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationsAndData.R
\name{ggplot_time_series}
\alias{ggplot_time_series}
\title{Plot time series simulations with experimental data}
\usage{
ggplot_time_series(
  simulations,
  experiments,
  nrow = NULL,
  ncol = NULL,
  plot.state = FALSE
)
}
\arguments{
\item{simulations}{list of simualtions as output from the simulator}

\item{experiments}{list of experiments}

\item{show.plot}{boolean variable. Set show.plot=TRUE to display plots
when running the funcion, FALSE otherwise}
}
\value{
list of plots with simulations and experimental data
}
\description{
This function plots simulations of time series experiments and plots them
against experimental data. The input in the provided experiments must differ
only in one vector component.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- write_and_compile(as_ode(m))
ex <- experiments(m,o)
s <- simulator.c(ex,o)
p0 <- values(m$Parameter)
y <- s(p0)
ggplot_time_series(y,ex)
}
