% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationsAndData.R
\name{ggplotTimeSeriesStates}
\alias{ggplotTimeSeriesStates}
\title{Plot time series simulation with state variables}
\usage{
ggplotTimeSeriesStates(simulations, experiments, nrow = NULL, ncol = NULL)
}
\arguments{
\item{simulations}{list of simualtions as output from the simulator}

\item{experiments}{list of experiments}

\item{show.plot}{boolean variable. Set show.plot=TRUE to display plots
when running the funcion, FALSE otherwise}
}
\value{
list of plots with simulations and experimental data
}
\description{
This function plots simulations of time series experiments and plots them
against experimental data. The input in the provided experiments must differ
only in one vector component.
}
