% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{globalSensitivity}
\alias{globalSensitivity}
\title{Global Sensitivity Analysis}
\usage{
globalSensitivity(parSample, outputSample, nBins = "Sturges")
}
\arguments{
\item{parSample}{a matrix of parameter vectors (rows)}

\item{outputSample}{a matrix, with rows of outputs (row-index is the sample index)}

\item{nBins}{number of bins, if unset defaults to the default of the hist function}
}
\value{
sensitivity S\link{i,j} of output\link{i} with respect to parameter\link{j}
}
\description{
This function performs a binning based estimation of the global
sensitivity of a model's output with respect to the model's
parameters. The output can be a prediction of the model's behaviour
in a scenario of interest (parameters, input, intial values,
boundary conditions, scheduled events etc.). The output models a
potentially measurable value (the "observable"). The sample-rows
and the output rows must correspond (they must be from the same
model simulation).
}
