% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{gradLog_NormalPrior}
\alias{gradLog_NormalPrior}
\title{Gradient of the logarithm of a normal prior}
\usage{
gradLog_NormalPrior(mean, sd)
}
\arguments{
\item{mean}{vector of mu values}

\item{sd}{vector of standard deviation values}
}
\value{
g(x) a closure that remembers mean and sd from its creation
}
\description{
This makes a function that returns grad(log(dprior(x))) The
returned function implictly remembers the parameters of the normal
distribution.
}
