% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{\%otherwise\%}
\alias{\%otherwise\%}
\title{This function can be used to specify default values}
\usage{
a \%otherwise\% b
}
\arguments{
\item{a}{value to check for NULL}

\item{b}{value to substitute}
}
\value{
a, or b if a is NULL
}
\description{
When attributes are missing, the \code{base::attr()} function returns
NULL. In those cases this function can be used to find an
alternative value in one expression:
\code{attr(x,"dim") \%otherwise\% length(x)}
}
