% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPosterior.R
\name{highCor}
\alias{highCor}
\title{highCor returns ordered index-pairs of high to low correlation}
\usage{
highCor(C)
}
\arguments{
\item{C}{the correlation matrix of a sample, no attributes need be
present other than dim.}
}
\value{
data.frame with columns i and j, representing the rows and columns of high to low correlation pairs.
}
\description{
This function uses the correlation matrix C of a sample X, orders
all values from the upper triangle of C (excluding the diagonal)
from highest to lowest correlation value and returns the indices as
a data.frame.
}
\details{
When truncated, the result can be used to plot only pairs with high
correlation.
}
\examples{
A <- matrix(
  c(
     1,  -1, 0.1,
    -1,   1, 0.4,
   0.1, 0.4,   1
  ),3,3
)
print(highCor(A))
}
