% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_SBtab.R
\name{import_experiments}
\alias{import_experiments}
\title{Reads the Data and Model Contained in an SBtab Document (tsv)}
\usage{
import_experiments(modelName = NULL, SBtabDir)
}
\arguments{
\item{modelName}{(string) the functions of the model have this prefix}

\item{SBtabDir}{(string) a local directory that contains tsv files (with SBtab content)}
}
\value{
list of simulation experiments (and the data corresponding to that simulation)
}
\description{
An SBtab Document is a set of tables that represent reactions,
compounds, parameters, and measured data that correspond to
simulations of the model under certain input conditions and initial
values.
}
\details{
This function assumes that this information is stored in a series
of tsv files. The content is imported using the SBtabVFGEN package.

The data contents are reorganized into a list of simulation
experiments (initial values, measurement time points, etc.)
}
