% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_SBtab.R
\name{import_from_SBtab}
\alias{import_from_SBtab}
\title{Import Systems Biology Models given in Tabular Form}
\usage{
import_from_SBtab(SBtabDir)
}
\arguments{
\item{SBtabDir}{the directory that contains `.tsv` files (with SBtab content)}
}
\value{
a model as a list of data.frames, one per tsv file, the
    Document title is attached as a comment attribute:
    comment(model) = Document Title
}
\description{
This function uses SBtabVFGEN to read a series of tsv files, each
containing one systems biology table that together create a model
(Reactions, Parameters, etc.). The model is converted to a vfgen
compatible file. This file is processed by vfgen through a system
call to create source code for R (deSolve) and C (GSL solvers).
}
\details{
This requires vfgen to be installed
(https://github.com/WarrenWeckesser/vfgen) - not an R package.

SBtab is a particular convention on how to structure the tables
(sbtab.net)
}
\examples{
 model <- import_from_SBtab("./model")
 comment(model)
 source("model.R")
}
