% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{kinetic_law_matrix}
\alias{kinetic_law_matrix}
\title{Split Kinetic Law}
\usage{
kinetic_law_matrix(r)
}
\arguments{
\item{r}{the reaction table (data.frame)}
}
\value{
a character matrix with a forward and backward column
}
\description{
This function performs a very simplified split of a kinetic law
into a forward part and a backward part, if it isn't pre-split in
the file.
}
\details{
If the data.frame contains separate forward and backward rates these will
be returned instead. Instead of using this function,
\code{m$Reaction[,c("fwd","bwd")]} would accomplish a very similar
thing.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAP79")) # not pre-split
print(colnames(m$Reaction))
k <- kinetic_law_matrix(m$Reaction)
print(k)
}
