% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{ll}
\alias{ll}
\title{Default Log-likelihood Function}
\usage{
ll(parMCMC)
}
\arguments{
\item{parMCMC}{a numeric vector, with attributes for MCMC, specifically smmala}
}
\value{
a scalar value: log(likelihood(data|parMCMC))
}
\description{
Extracts the \code{logLikelihood} value from the simulations attribute
of the parMCMC argument, requires:
\itemize{
\item parMCMC has simulations attribute
\item simulations list includes logLikelihood values (omit<3)
}
}
\details{
This function will take the log-likelihood-values calculated by the
ode solver in this package, and return the sum of those values over
all experiments. The value the simulator returns is calculated
with the assumption of a normal distribution on measurement errors.

This function does \emph{almost no work}, it merely sums up the values
calculated during simulation.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
c_path(o) <- write_c_code(generate_code(o))
so_path(o) <- shlib(o)
ex <- experiments(m,o)
s <- simulator.c(ex,o,omit=2) # not 3
p <- values(m$Parameter)
attr(p,"simulations") <- s(p)
print(ll(p))
}
