% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{loadODE}
\alias{loadODE}
\title{Load an ODE model from a file}
\usage{
loadODE(fileList)
}
\arguments{
\item{fileList}{list of file paths}
}
\value{
a list of data.frames describing the model
}
\description{
The input can be either a list of files, or one compressed archive
file like "model.tar.gz". This function determines the name of the
model from the name of the archive, or from the parent directory of
uncompressed tsv files. This auto-determined name is attached as a
comment of the return value. To override this choice, replace the
comment. This function exists for compatibility with the scripts in
the RPN-derivative repository and loads the zip and tar.gz files
that SBtabVFGEN::sbtab_to_vfgen writes to disk.
}
