% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{loadSubSample_mpi}
\alias{loadSubSample_mpi}
\title{This function merges mpi-samples into one}
\usage{
loadSubSample_mpi(files, size = NA, selection = NA)
}
\arguments{
\item{files}{the files where the individual samples are stored}
}
\value{
a list of matrices, by temperature, concatenated.
}
\description{
When using MPI, we save the sample immediately into a file, each
rank saves to its own file. This function collects all of these
smaller samples into one. The samples should be saved with
\code{saveRDS()}.
}
