% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{log10ParMapJac}
\alias{log10ParMapJac}
\title{LOG10 parameter mapping, jacobian}
\usage{
log10ParMapJac(parMCMC)
}
\arguments{
\item{parMCMC}{the sampling variables (numeric vector)}
}
\value{
a numeric matrix (dim: \code{c(length(parMCMC),length(parMCMC))})
}
\description{
This map is used by the simulator to transform sampling variables
into ODE-model parameters. As we often calculate sensitivities, we
also need the Jacobian of the map, due to the chain rule of
differentiation.
}
\examples{
p <- c(-1,0,1)
parMap <- log10ParMap
parMpJ <- log10ParMapJac
print(parMap(p))
print(parMpJ(p))
}
