% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Default log-likelihood function}
\usage{
logLikelihood(experiments)
}
\arguments{
\item{experiment}{will be compared tp the simulation results}
}
\description{
This returns a function f(simulations), which maps simulation
results to log(likelihood) values. The experiments are used
implicitly; simulations is a list as returned by
rgsl::r_gsl_odeiv2_outer().
}
