% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{logParMap}
\alias{logParMap}
\title{NATURAL LOG parameter mapping used by the MCMC module}
\usage{
logParMap(parMCMC)
}
\arguments{
\item{parMCMC}{the sampling variables (numeric vector)}
}
\value{
a numeric vector intended for the simulator.
}
\description{
This map is used by the simulator to transform sampling variables
into ODE-model parameters.
}
\examples{
p <- c(-1,0,1)
parMap <- logParMap
print(parMap(p))
}
