% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{logParMapJac}
\alias{logParMapJac}
\title{NATURAL LOG parameter mapping, jacobian}
\usage{
logParMapJac(parMCMC)
}
\arguments{
\item{parMCMC}{the sampling variables (numeric vector)}
}
\value{
a numeric matrix (dim: \code{c(length(parMCMC),length(parMCMC))}).
}
\description{
This map is used by the simulator to transform sampling variables
into ODE-model porameters. As we often calculate sensitivites, we
alos need the jacobian of the map, due to the chain rule of
differentiation.
}
\examples{
p <- c(-1,0,1)
parMap <- logParMap
parMpJ <- logParMapJac
print(parMap(p))
print(parMpJ(p))
}
