% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{makeAcceptanceProbability}
\alias{makeAcceptanceProbability}
\title{Ths function Creates an acceptanceProbability function}
\usage{
makeAcceptanceProbability(
  experiments,
  modelName,
  getAcceptanceProbability = defaultAcceptance,
  parMap = identity,
  simulate = NULL
)
}
\arguments{
\item{experiments}{a list of experiments}

\item{modelName}{an annotated string, with the model name and model file as comment}

\item{getAcceptanceProbability}{an R function that mape the results of a simulation and experimental data to an acceptance probability}

\item{parMap}{an optional mapping between sampling parameters (parABC) and model parameters (e.g. rescaling,re-ordering).}
}
\value{
a function that calculates probabilities given only parABC as input; it implicitly uses all the argiments to this function.
}
\description{
The returned closure needs only the sampling variables (parABC) as
inputand calculates a probability of accepting Markov chainmoves.
}
