% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{makeObjectiveSSA}
\alias{makeObjectiveSSA}
\title{Function that creates the objective function}
\usage{
makeObjectiveSSA(
  experiments,
  model.tab,
  parNames,
  distance,
  parMap = identity,
  outputFunction = identity,
  vol = 4e-16,
  unit = 1e-06,
  reactions,
  nStochSim = 1
)
}
\arguments{
\item{experiments}{a list of experiments.}

\item{model.tab}{an SBtab model.}

\item{parNames}{the names of the (biological) parameters of the
model.}

\item{distance}{a user supplied function that calculates a distance
between simulation and data with an interface of
distance(simulation, data, errVal), where errVal is an estimate
of the measurement noise (e.g. standard deviation), if needed
by the function.}

\item{parMap}{a function that translates ABC variables (parABC)
into something the model will accept.}

\item{outputFunction}{a function that, given (t,state,param), outputs the quantity that is measured in the experiments (e.g., the outputFunction may output the concentration of one of the compounds in the system).}

\item{vol}{Volume in which the reactions take place.}

\item{unit}{unit of measure for the volume (it converts the volume into liters).}

\item{reactions}{a list of GillespieSSA2 reactions}

\item{nStochSim}{number of stochastic simulations to average over.}
}
\value{
a closure for the objective function that implicitly
depends on all of the arguments to this function but explicitly
only on the ABC parameters parABC.
}
\description{
Given a parameter set, this function computes the distance between
experimental data and simulated data (coresponding to the parameter
in input).
}
