% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{match.coefficients}
\alias{match.coefficients}
\title{find the coefficients in a formula}
\usage{
match.coefficients(chrv)
}
\arguments{
\item{chrv}{a character vector as returned by parse.formula}
}
\value{
coefficients, as a vector
}
\description{
A reaction formula has reactants and products, separated by <=>,
with reactants on the left and products on the right (by convention).
Each of those is a plus separated list of reacting compounds and
modifiers, with optional coefficients, e.g.: A + 2*B <=> AB2
}
\details{
Once the formula is split into left and right side, this function
determines the coefficients. For the above example, this function
returns c(1,2) for the left side and 1 for the right side.
}
