% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{method}
\alias{method}
\title{Find Integer}
\usage{
method(name)
}
\arguments{
\item{name}{character scalar, name of the method}
}
\value{
an integer that is acceptable to \link{simfi} and \link{simulator.c}
}
\description{
Given a ODE solver name (from the GSL solver module odeiv2), return
an integer offset {0..10}.
}
