% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{metropolis_update}
\alias{metropolis_update}
\title{Metropolis Update is an MCMC update function}
\usage{
metropolis_update(
  simulate,
  logLikelihood = ll,
  dprior = function(x) prod(dnorm(x)),
  Sigma = NULL,
  parAcceptable = function(p) {
     all(is.finite(p))
 }
)
}
\arguments{
\item{simulate}{a function that simulates the model}

\item{logLikelihood}{a function that returns the log-likelihood
value given the parameter value, with simulations attached to
the parameter as an attribute (probably a closure)}

\item{dprior}{a function that returns the prior density of the
given parameter vector}

\item{Sigma}{the transition kernel's covariance matrix.}

\item{parAcceptable}{a function that can be used to reject a
proposal based on the values of the parameters alone (shortcut
to rejection, sans simulation)}
}
\value{
a closure with the arguments: (parGiven, eps=1e-4), where
\code{parGiven} is the current position of the Markov chain (a
numeric vector, with attributes), and \code{eps} the current step
size.
}
\description{
During Markov chain Monte Carlo a given parameter needs to be
updated, the model needs to be simulated at the updated point.
}
\details{
Using the simulations, and an acceptance rule, the proposed update
is either accepted or rejected.

This function returns a closure \code{metropolis}, with only \code{parMCMC}
as it's sole argument: \code{parProposal <- metropolis(parGiven)}

An optional argument to this function is \code{parAcceptable}, during
sampling, when \code{metropolis} is called as the update function, and
\code{parAcceptable(parProposal)} returns \code{FALSE}, then metropolis
shortcuts to \code{return(parGiven)} without performing simulations.

This function can be used to weed out parameter combinations that
would result in obviously nonsensical simulations without wasting
CPU-time.

The return value is a function (closure) that operates on mcmc
variables, these variables are numeric vectors with some necessary
attributes. The attributes record this vector's simulation results,
and other derived quantities.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
c_path(o) <- write_c_code(generate_code(o))
so_path(o) <- shlib(o)
ex <- experiments(m,o)
s <- simulator.c(ex,o,omit=0)
dprior <- dNormalPrior(values(m$Parameter),m$Parameter$stdv)
p <- mcmc_init(1.0,values(m$Parameter),s,ll,dprior)
UP <- metropolis_update(s,ll,dprior=dprior,Sigma=diag(m$Parameter$stdv)^2)
p2 <- UP(p)
## updated value:
print(p2)
print(sum(abs(p2-p)))
}
