% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{moleCountConversion}
\alias{moleCountConversion}
\title{This function calculates the conversion from moles to particle counts}
\usage{
moleCountConversion(unit)
}
\arguments{
\item{unit}{character vector, will be parsed for SI prefixes and to
determine the mole component}
}
\value{
a data.frame with an lvpower and factor component, with
names like the names of the unit vector
}
\description{
Given some molar quantity with a unit (character vector), this function
calculates a conversion factor based on the SI prefixes used, as
well as what the exponent of LV (Avogadro's constant * Volume)
applies in this case.
}
