% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{name_method}
\alias{name_method}
\title{Reverse look-up of method name from key}
\usage{
name_method(key = seq(0, 10))
}
\arguments{
\item{key}{an integer from 0 to 10 (this is used as an offset in c, for 11 items)}
}
\value{
a string representation of the integration method.
}
\description{
These are the methods in the gsl library (documented in the
official documentation), but in reverse order, as they are
approximately ordered by complexity, with more complex methods
usually being better (but slower).
}
\details{
It is therefore a reasonable approach to try methods from the more
complex end of the list first and try the next method if the
solutions are too slow. But we need to check the accuracy/stability of the
result. The mapping between method names and keys:

\if{html}{\out{<div class="sourceCode">}}\preformatted{     msbdf:	 0
   msadams:	 1
     bsimp:	 2
    rk4imp:	 3
    rk2imp:	 4
    rk1imp:	 5
     rk8pd:	 6
      rkck:	 7
     rkf45:	 8
       rk4:	 9
       rk2:	10
}\if{html}{\out{</div>}}

The returned value is an integer index.
}
\examples{
print(name_method())
}
