% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{observable.mean.in.bin}
\alias{observable.mean.in.bin}
\title{The mean value of an observable value given a parameter bin}
\usage{
observable.mean.in.bin(id, outputSample)
}
\arguments{
\item{id}{is an integer vector that identifies the bin the
parameter vector falls into to create the same row of the
outputSample (the output stems from a model simulation with
parameters). id has the same length as outputSample rows.}

\item{outputSample}{a matrix of output values, one output vector
per row (different rows are results at different parameter
values)}
}
\value{
M\link{i,j} the mean of each observable\link{j} in bin\link{i}
}
\description{
The model parameters were binned with a histogram function. In each
bin one of the parameters is almost fixed (it varies much less than
the other parameters \link{full range}). This function returns the mean
of the observable for each bin, as a vector.
}
