% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{onlyCoefficients}
\alias{onlyCoefficients}
\title{Returns a list of reaction coefficients}
\usage{
onlyCoefficients(formulaList)
}
\arguments{
\item{formulaList}{a list of character vectors, derived from the
left or right side of a reaction formula:}
}
\value{
a list of numeric coefficient vectors
}
\description{
This function maps c("A","2 B") to c(1,2)
}
\details{
This is a C function because it is much easier to write in C.  C
has the strtod() function which expects a leading number and stops
when the numbers end. as.character() returns NA if the input
contains any dirt.

The reaction formula is as string like this: "A + 2 B <=> C", when
split at \verb{<=>} and then later at \code{+}, we get the strings that must
be parsed: "A" and "2 B" for the left side and "C" for the right
side. The numbers are the stoichiometric constants, or coefficients.
}
\examples{
print(onlyCoefficients("12 A"))
}
