% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{onlyNames}
\alias{onlyNames}
\title{Returns only the names in a reaction formula}
\usage{
onlyNames(formulaList)
}
\arguments{
\item{formulaList}{a list of strings like: "2 B" or "45 X"}
}
\value{
a list of name vectors
}
\description{
This is the companion function to onlyCoefficients. It returns the
names of reactants, without the stoichiometry.
}
