% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{parameter.from.kinetic.law}
\alias{parameter.from.kinetic.law}
\title{Attempt to find multiplicative reaction rate coefficients}
\usage{
parameter.from.kinetic.law(kineticLaw, tab)
}
\arguments{
\item{kineticLaw}{a string with a mathematical formula}

\item{tab}{an SBtab document, as returned by \code{SBtabVFGEN::sbtab_from_tsv()}}
}
\value{
a parameter with value, and unit as attribute
}
\description{
This function assumes that the kinetic Law is mass action
kinetics. This function helps in converting the units of an ODE
into units that work in stochastic simulations. Converting the
units of a general formula (Michaelis Menten, Hill kinetics, etc.)
is difficult and dubious in stochastic simulations.
}
\details{
For these reasons, this functions assumes: \verb{kf*A*B*[...]} with the
first word \code{kf} representing the reaction rate coefficient.

Given an SBtab document, this function finds the value (if any) and
unit of this coefficient. The coefficient itself can be defined as
a fixed constant, a parameter, or an algebraic expression in the
document. The most important attribute here is the unit.
}
