% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{parameterReaction}
\alias{parameterReaction}
\title{Find the Reaction a parameter appears in}
\usage{
parameterReaction(reactionKinetic, parNames)
}
\arguments{
\item{reactionKinetic}{a character vector with all of the kinetic
law expressions}

\item{parNames}{parameter names}
}
\value{
list with two named components: fwd, bwd; each is a named
vector of integers, th enames are taken from parNames, the
integer indicates a reaction.
}
\description{
Usually Parameters are a list of names and values, Reactions have
reaction kinetics, which can contain any number of parameters. To
make matters worse, a parameter may influence two or more
reactions. The goal is to decide on the parameter's conversion
factor. We shall assume that if the parameter appears in more than
one reaction, it is the same kind of context, and thus it must be
converted exactly the sam eway.  This function finds the first
reaction a given parameter appears in.
}
\details{
Note that the entire conversion may not work at all for complicated
kinetics, we use very simple rules here.
}
