% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{parameters_from_expressions_func}
\alias{parameters_from_expressions_func}
\title{Function that creates a function that reads the SBtab expression table and computes the expressions (which are functions of parameters) given the parameters passed in input}
\usage{
parameters_from_expressions_func(model.tab)
}
\arguments{
\item{model.tab}{an SBtab file imported in R, as it is returned by the function SBtabVFGEN::sbtab_from_tsv}
}
\value{
a function that given a vector of (sampling) parameters returns a vector of parameters corresponding to the evaluations of the expressions in the SBtab expression table (which are functions of the parameters)
}
\description{
Function that creates a function that reads the SBtab expression table and computes the expressions (which are functions of parameters) given the parameters passed in input
}
