% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{parse.formula}
\alias{parse.formula}
\title{Splits a formula into a left and right side}
\usage{
parse.formula(reactionFormula)
}
\arguments{
\item{string}{\itemize{
\item reactionFormula
}}
}
\value{
a named list with a forward component and a backward
component, each entry contains a character vector
}
\description{
This function splits a reaction formulka apart into its parts,
removing whitespace on each side: "A + 2\emph{B <=> AB2" will be split
into a list with two entries
list$reactants == c("A","2}B")
list$products == c("AB2")
}
