% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{parse.kinetic}
\alias{parse.kinetic}
\title{Find forward and backward component in a reaction kinetic}
\usage{
parse.kinetic(reactionKinetic)
}
\description{
a reaction kinetic can be almost any function, and in general it is
not possible to tell apart which part of a kinetic law is which.
}
\details{
But for mass action kinetics, and positive reaction rate
coefficients, the expressions mostly look like this:

kf*prod(reactants.concentration) - kb*prod(product.concentrations)

this functions splits at '-', and if none is present, then the
reaction is assumed to be irreversible.

In a more general setting (where the '-' split is wrong),
the splitting has to be done by hand or more complex rules.
}
