% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concise.R
\name{parse_concise}
\alias{parse_concise}
\title{Read Concise Error Notation}
\usage{
parse_concise(
  v,
  use.errors = requireNamespace("errors", quietly = TRUE),
  na = c(NA, NA)
)
}
\arguments{
\item{v}{a character vector of numbers in concise error notation}

\item{use.errors}{if TRUE, the errors package will be used to
return an object of type "errors" (from that
package). Otherwise, the errors will be attached as an
attribute (also called "errors" to be consistent with the
errors package)}

\item{na}{a two element vector which will replace NA values,
e.g. c(NA,NA); na=c(0,Inf) means \emph{infinite uncertainty} for missing values}
}
\value{
either a numeric object with class errors (with the same
dimensions as \code{v}), or a numeric matrix of values and
uncertainties (2 rows), dimensions of original object are lost
}
\description{
Convert a vector of strings of the form: c("1.2(3)E-4","1.2(3)E-2")
to a matrix with two rows:
\enumerate{
\item values,
\item uncertainties.
}
}
\details{
If the errors package is available, then an \emph{errors} object is
returned instead (uncertainties are an attribute). In that case the
dimensions of \code{v} are preserved on output. You can override this
choice using the second argument \code{use.errors}.

Concise error notation means that a floating point number is
followed by an integer in parentheses which indicates the
uncertainty of the last digits of the value:
\deqn{1.2345(12) = 1.2345 \pm 0.0012}{1.2345(12) = 1.2345 ± 0.0012}.

If the errors package is installed, then it will be used to
represent the return value.
}
\examples{
x <- parse_concise(c("1.23(4)","0.51099895069(16)","1.25663706127(20)e-6","1.3±1.6","5;1"))
print(as.data.frame(x))
}
