% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{parUpdate}
\alias{parUpdate}
\title{Updates Parameter Values}
\usage{
parUpdate(objectiveFunction, curPar, canPar, curDelta, curPrior, delta, dprior)
}
\arguments{
\item{objectiveFunction}{function that, given a (vectorial)
parameter as input, simulates the model, and outputs the
distance between experimental data and data simulated from the
model with the parameter provided in input}

\item{curPar}{current parameter values (as ABC samples them)}

\item{canPar}{candidate parameter values (for MCMC)}

\item{curDelta}{current distance between data and simulation, if
the MCMC chain has not yet reached any point where this is
below the threshold (delta), this can be accepted as the new
current state for the chain.}

\item{curPrior}{current Prior values given curPar}

\item{delta}{distance threshold for ABC}

\item{dprior}{prior probability density function}
}
\value{
updated values for curPar, curDelta, and curPrior
}
\description{
under valid ABC update conditions (successful simulation) the
parameters are updated to new values.
}
