% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationsAndData.R
\name{plot_time_series_base}
\alias{plot_time_series_base}
\title{plot function for experiments}
\usage{
plot_time_series_base(simulations, experiments, by = 1, ylimit = NULL)
}
\arguments{
\item{simulations}{simulation results (a list)}

\item{experiments}{experiment setup (a list)}

\item{by}{skip this many sampled lines in between plotted lines}

\item{ylimit}{y-axis limits}
}
\value{
plot object
}
\description{
This function does not use ggplot2, only base plot functions like
lines() and arrows().
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- write_and_compile(as_ode(m))
ex <- experiments(m,o)
s <- simulator.c(ex,o)
p0 <- values(m$Parameter)
y <- s(p0)
plot_time_series_base(y,ex)
}
