% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationsAndData.R
\name{plotDoseResponse}
\alias{plotDoseResponse}
\title{Plot dose response simulations with experimental data}
\usage{
plotDoseResponse(simulations, experiments, dose, show.plot = TRUE)
}
\arguments{
\item{simulations}{list of simualtions as output from the simulator}

\item{experiments}{list of experimental data from the same dose response experiment}

\item{dose}{vector of dose values to plot on the x axis}

\item{show.plot}{boolean variable. Set show.plot=TRUE to display plots
when running the funcion, FALSE otherwise}
}
\value{
plot with simulations and experimental data
}
\description{
This function plots simulations of one dose response experiment and plots them
against experimental data.
}
