% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulations.R
\name{plotSimualtionsFromSBtab}
\alias{plotSimualtionsFromSBtab}
\title{Simulate and plot Data and Simulation}
\usage{
plotSimualtionsFromSBtab(
  SBtabDir,
  paramVal,
  plotDir = NULL,
  width = 15,
  heigth = 10
)
}
\arguments{
\item{SBtabDir}{the directory that contains `.tsv` files (with SBtab content)}

\item{paramVal}{parameter vector with which the model has to be simulated}

\item{plotDir}{the directory where the plots will be saved as .pdf and .RData variables}

\item{width}{width of the plot window}

\item{heigth}{heigth of the plot window}
}
\value{
a vector of R plot objects
}
\description{
This function imports model and experimental data saved as SBtab files,
it simualtes the model with the same initial conditions and input as in
the corresponding experiments, and it plots the simulations together with
the corresponding experimental data.
It currently works with one dimensional output.
}
