% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationsAndData.R
\name{plotTimeSeriesBase}
\alias{plotTimeSeriesBase}
\title{plot function for experiments}
\usage{
plotTimeSeriesBase(
  simulations,
  experiments,
  nmax = NULL,
  by = 1,
  ylimit = NULL
)
}
\arguments{
\item{simulations}{simulation results (a list)}

\item{experiments}{experiment setup (a list)}

\item{nmax}{maximum index of lines to plot from sample, between 1 and size of sample}

\item{by}{skip this many sampled lines in between plotted lines}

\item{ylimit}{y-axis limits}
}
\value{
plot object
}
\description{
This function does not use ggplot2, only base plot functions like
lines() and arrows().
}
