% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{print.experiments}
\alias{print.experiments}
\title{prints the simulation experiments}
\usage{
\method{print}{experiments}(x, ...)
}
\arguments{
\item{x}{simulation experiments with data}

\item{...}{ignored.}
}
\value{
called for side-effect (printout); no value.
}
\description{
The experiments, if accidentally printed, are difficult to read.
This function prevents these accidental prints. It summarizes the
data and simulation experiments instead.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
ex <- experiments(m,o)
print(ex)
}
