% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{print.ode}
\alias{print.ode}
\title{Print a summary about the ode}
\usage{
\method{print}{ode}(x, ...)
}
\arguments{
\item{x}{the ode}

\item{...}{requirement of print generic, not used.}
}
\value{
called for side-effect (printout); no value.
}
\description{
An ODE model was crteated by \code{as_ode} can be summarized here,
including information about the compiled version of the model.
}
\details{
The ode model is for the most part a list of named vectors and
matrices which together encode the mathematical structure of the
ode.
}
\examples{
f <- uqsa_example("AKAR4")
m <- model_from_tsv(f)
o <- as_ode(m)
print(o)
}
