% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{propensity}
\alias{propensity}
\title{propensity creates a propensity formula}
\usage{
propensity(conv.coeff, kinetic.law, rExpressions)
}
\arguments{
\item{conv.coeff}{propensity conversion coefficient: conv.coeff*kinetic.law = propensity function}

\item{kinetic.law}{the kinetic law of this reaction (as used with ODEs)}

\item{rExpressions}{named math expressions that appear in the kinetic.law of this reaction}
}
\value{
a string representation of the propensity function
}
\description{
given the custom math expressions needed to calculate a propensity,
the propensity coefficient and the kinetic law of the reaction,
this function makes a string that can be used with GillespieSSA2.
}
\details{
The propensity coefficient translates between
}
