% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{reaction<-}
\alias{reaction<-}
\title{Add a Reaction to an ODE}
\usage{
reaction(vf, r, p) <- value
}
\arguments{
\item{vf}{a named character vector of the right length (number of state variables)}

\item{r}{a named vector of stoichiometric coefficients for the reactants}

\item{p}{a named vector of stoichiometric coefficients for the products}
}
\value{
updated vf
}
\description{
This function operates on an ODE vector field. It makes it possible
to construct the right hand side one reaction at a time:
\code{reaction(vf,r,p) <- flux} will add a flux term to the ODE's right
hand side (vector field), adding the given \code{flux} (reaction
kinetic) in the right places, informed by the reactants and
products.
}
\details{
The reactants, e.g.: \code{c(A=2,B=1)} name which state variables are affected by the flux (negatively, for reactants)
The products, e.g.: \code{c(C=1)} name which state variables are affected positively by the flux.
}
\examples{
Reaction <- "A + B <=> C"
r <- c(A=1,B=1)
p <- c(C=1)
vf <- c(A="",B="",C="") # empty
reaction(vf,r,p) <- "A*B-C"
}
