% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{replace_powers}
\alias{replace_powers}
\title{replace_powers does string manipulation}
\usage{
replace_powers(v)
}
\arguments{
\item{v}{a character vector}
}
\value{
a string where all occurrences of \code{^} have been replaced by function calls like \code{pow()}
}
\description{
This function takes a string argument with human readable math
(e.g. R code), and replaces the power operator \code{z^n} with
C-compatible function calls: \code{pow(x,n)}, it counts parentheses to
determine the base and exponent automatically.
}
\details{
This functions assumes that gsl functions can be used, the GNU
Scientific Library includes powers of small integers. These
functions may be faster than always calling \code{pow} from \code{math.h}.

This is necessary because in C the \code{^} operator means something
else (exclusive bitwise xor for integers). No attempt will be made
to cast the numbers to \code{float} or \code{double}.
}
\examples{
print(replace_powers(c("2^3.1","10^-6","x^2","(1+(1+x))^(n-0.5)")))
}
