% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{Rmpi_swap_temperatures}
\alias{Rmpi_swap_temperatures}
\title{Communicate with other ranks and swap beta}
\usage{
Rmpi_swap_temperatures(i, B, LL, H, r, comm, cs)
}
\arguments{
\item{i}{MCMC iteration}

\item{B}{inverse temperature (parallel tempering)}

\item{LL}{log-likelihood value of current point}

\item{H}{algorithm's step size (often called epsilon in literature)}

\item{r}{MPI rank}

\item{comm}{MPI communicator}

\item{cs}{MPI comm size}
}
\description{
Given a current log-likelihood, temperature and step-size, this
decides whether to send or receive the same variables from a
neighboring process and swap temperatures with them.
}
