% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{rNormalPrior}
\alias{rNormalPrior}
\title{rNormalPrior returns a random vector generator}
\usage{
rNormalPrior(mean, sd)
}
\arguments{
\item{mean}{mean of the random variables (a vector)}

\item{sd}{standard deviation of the random variables (same size vector as
mean)}
}
\value{
an independent multivariate normal random vector generating function: rprior(n),
where n is the requested number of vectors (rows)
}
\description{
The return value is a function that generates random vectors of the
same size as mean and sd from a multivariate normal distribution with
independent components with mean "mean" and standard deviation "sd".
The random vectors are returned as n
rows of a matrix, where n is the only argument of the returned
function.
}
\examples{
rnp<-rNormalPrior(mean=c(0,1,2),sd=c(1,2,3))
rnp(12)
}
