% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{rUniformPrior}
\alias{rUniformPrior}
\title{rUniformPrior returns a random vector generator}
\usage{
rUniformPrior(ll, ul)
}
\arguments{
\item{ll}{lower limit of the random variables (a vector)}

\item{ul}{upper limit of the random variables (same size vector as
ll)}
}
\value{
a uniform random vector generating function: runiform(n),
    where n is the requested number of vectors (rows)
}
\description{
The return value is a function that generates random vectors of the
same size as ll and ul from a uniform distribution within the
limits defined by ul and ll. The random vectors are returned as n
rows of a matrix, where n is the only argument of the returned
function.
}
\examples{
rup<-rUniformPrior(ll=c(0,1,2),ul=c(1,2,3))
rup(12)
          [,1]     [,2]     [,3]
 [1,] 0.7879692 1.344263 2.452547
 [2,] 0.7154937 1.846704 2.174703
 [3,] 0.6644920 1.257256 2.764870
 [4,] 0.3470803 1.524026 2.715229
 [5,] 0.7879692 1.344263 2.452547
 [6,] 0.7154937 1.846704 2.174703
 [7,] 0.6644920 1.257256 2.764870
 [8,] 0.3470803 1.524026 2.715229
 [9,] 0.7879692 1.344263 2.452547
[10,] 0.7154937 1.846704 2.174703
[11,] 0.6644920 1.257256 2.764870
[12,] 0.3470803 1.524026 2.715229
}
