% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{saltelli_prior}
\alias{saltelli_prior}
\title{Outpts the random sample on which to perform the Sobol-Homma-Saltelli global sensitivity
analysis}
\usage{
saltelli_prior(nSamples, rprior)
}
\arguments{
\item{nSamples}{number of rows to return}

\item{rprior}{a function that samples from the prior distribution}
}
\value{
a list with the components M1, M2 (both matrices) and N (a
3D-array).
}
\description{
Each parameter vector has length nPars,
The sample consists of two random (nSamples x nPars) matrices M1,
M2 and a third (nSamples x nPars x nPars) array N. N consists of
nPars copies of M2, except that in each M2-matrix one column has
been replaced by the corresponding column of M1. M1 and M2 consists
of random numbers from a normal distribution.
}
\details{
These matrices provide prior distribution samples to be further
processed by the simulator, similar to this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{sim <- simulator.c(experiments,modelName)
fM1 <- t(sim(t(M1))[[1]]$state[,ti,])       # or similar
}\if{html}{\out{</div>}}

For details see: Halnes, Geir, et al. J. comp. neuroscience 27.3 (2009): 471.
}
