% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaFunctions.R
\name{sampleWithNoise}
\alias{sampleWithNoise}
\title{(for testing) A non-Copula sampling function as fallback}
\usage{
sampleWithNoise(X, sdf = 0.01, ...)
}
\arguments{
\item{X}{an N×M matrix (N is the sample size), M is the number of variables (MCMC or ABC vars)}

\item{sdf}{factor to increase or decrease the standard deviation of the added noise}

\item{...}{passed to \code{base::sample.int()}}

\item{size}{size of returned sample (passed to \code{sample.int()})}
}
\value{
a matrix with \code{size} rows and \code{nrow(X)} columns.
}
\description{
If the sample is not suited to infer a Copula (fitCopula fails),
this fuction uses much simpler rules to re-draw a new sample from
an older sample with some added noise.
}
\details{
This can be used during testing, in cases where the acceptance was
very low and we have to deal with a very low quality sample This
function should work like \code{base::sample}, but adds small
noise. Missing values are always removed.
}
