% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{scaleParameter}
\alias{scaleParameter}
\title{Retruns a string that contains code to scale the parameter}
\usage{
scaleParameter(f, x, name, arg = "par")
}
\arguments{
\item{f}{the value of the scaling factor (a numeric vector), derived from the unit of the knietic parameter and stoichiometry.}

\item{x}{the reaction order based exponent of LV}

\item{name}{the name of the parameter}
}
\value{
string with scaling instructions (C)
}
\description{
The goal is to convertthe kinetic parameter k to the stochastic
parameter c.
}
\details{
The input is a pre-calculated factor f, determined from the
parameter's unit and reaction order. LV is an importantfactor for
conversion between molecule numbers and concentrations, the
appropriate exponent for LV is passed as x: c = f * LV^x * k
}
