% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivityEquilibriumApproximation}
\alias{sensitivityEquilibriumApproximation}
\title{Equilibrium state approximation of the solution sensitivity for ODE systems}
\usage{
sensitivityEquilibriumApproximation(experiments, model, parMap = identity)
}
\arguments{
\item{experiments}{a list of simulation experiments}

\item{model}{a list of functions for the model the experiments are applicable to}

\item{parMap}{a map to transform parMCMC into p, parameters the model accepts}

\item{simulations}{an equivalent list of simulation results, for one parameter vector}

\item{parMCMC}{the parameters that are used in Markov chain Monte Carlo as the MC variable}
}
\value{
S, the state sensitivity matrix length(x) × length(p) × length(t)
}
\description{
In this context, the sensitivity S(t;x,p) is dx(t;p)/dp, where
x(t;p) is the parameterized solution to an initial value problem
for ordinary differential equations and t is the independent
varibale: x'=f(t,x;p), where «'» indicates the derivative with
respect to t. In cases where you have a proxy variable for p,
e.g. r=log(p), the chain rule applies. Similarly, we also have an
output sensitivity for the function g(x(t;p)).  The equilibrium
approximation is exact for state-variable values close to an
equilibrium point q(p) (fixed-point): f(t,q(p);p)=0.
}
\details{
Typically, the sensitivity needs to be known at different
time-points t_k, therefore, this function returns S as a
3-dimensional array S[i,j,k], where the index k corrsponds to time
t_k; the closer x(t_k) is to equilibrium, the better the
approximation; near the initial state, the sensitivity is also
correct (only the intermediate time-span is approximate).

This function requires pracma::expm to work.
}
