% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{shlib}
\alias{shlib}
\title{Compile C code to shared library}
\usage{
shlib(file, verbose = getOption("uqsa.verbose", default = interactive()))
}
\arguments{
\item{file}{the c file that is to be compiled, OR an ODE/CME object
with a c.file defined and recorded in it.}

\item{verbose}{print decition outcomes about the compiler options}
}
\value{
the path of the created shared library
}
\description{
Calls \verb{R CMD SHLIB} to create the model's shared library.
}
\details{
The first argument can be a raw character scalar with just the path
of the c code to be compiled, or alternatively an object that has
this information stored within it. The models returned by \link{as_cme}
and \link{as_ode} can both carry this information, attach it via
\link{c_path<-}.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
C <- generate_code(o)
c_path(o) <- write_c_code(C)
so_path(o) <- shlib(o)
print(o)
if (file.exists(so_path(o))) cat("shared library exists.\n")
}
