% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{shs_gsa}
\alias{shs_gsa}
\title{Outputs the global sensitivity scores SI and SIT, calculated by the Sobol-Homma-Saltelli method}
\usage{
shs_gsa(fM1, fM2, fN, subtractMean = TRUE)
}
\arguments{
\item{fM1}{output (f)unction values for M1, nSamples × nOuts}

\item{fM2}{output (f)unction values for M2, nSamples × nOuts}

\item{fN}{output (f)unction values for N, nSamples × nOuts × nPars}
}
\value{
a list with sensitivity indices $SI and total sensitivities $SIT
}
\description{
M1, M2, and N are matrices prepared by \code{uqsa::shs_prior()}.  The
parameters (rows) from these matrices need to be simulated (using
any method), to obtain fM1, fM2 and fN.
}
\details{
These matrices are shaped similarly to M1, M2 and N respectively,
but now the parameters are replaced by the effects they have on a
observable of interest (the output). It can be the vector
valued output at a specific (single) time-point or a scalar output
at different time-points.

See Geir Halnes et al. (Halnes, Geir, et al. J. comp. neuroscience 27.3 (2009): 471.
}
