% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{simfiGaussianGradLogLikelihood}
\alias{simfiGaussianGradLogLikelihood}
\title{\link{SMMALA} Extract the gradient of the log-likelihood from the simfi solver's return value}
\usage{
simfiGaussianGradLogLikelihood(
  ParMapJac = function(x) {
     diag(1, length(x))
 }
)
}
\arguments{
\item{ParMapJac}{Jacobian of the parMap function}
}
\value{
the gradient of the Gaussian log-likelihood with respect to
the MCMC variable
}
\description{
This function extracts the approximate gradient of the log-likelihood and
transforms the gradient using the Jacobian of the parameter map
between Markov chain variables and model parameters.
}
\details{
The simfi() gradient is with respect to the raw model parameters
}
