% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{simfiGaussianLogLikelihood}
\alias{simfiGaussianLogLikelihood}
\title{\link{SMMALA} The default Extractor of the log-likelihood computed bu the simfi solver}
\usage{
simfiGaussianLogLikelihood(init = 0)
}
\arguments{
\item{init}{a base value that will be added to the log-likelihood}
}
\value{
a log-likelihood value for the set of experiments the
solver was set up with.
}
\description{
This function will only extract the log-likelihood value from the
solution via simfi. Simfi returns the likelihhod value based on the
assumption that the data provided in the list of experiments has a
Gaussian standard error. The value includes the normalising factor 1/sqrt(2\emph{pi}sigma^2)
for each measured value (the logarithm).
}
