% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{simulateAndComputeDistance}
\alias{simulateAndComputeDistance}
\title{Function that simulates a stochastic trajectory with the Gillespie algorithm
given certain experimental conditions and a parameter vector, and computes the
distance between the simulation and the experimental data}
\usage{
simulateAndComputeDistance(
  e,
  param,
  parMap = parMap,
  Phi = Phi,
  parameters_from_expressions = parameters_from_expressions,
  nStochSim = nStochSim,
  reactions = compiled_reactions,
  modelName = modelName,
  distance = distance
)
}
\arguments{
\item{e}{an experiment}

\item{param}{a named parameter vector}

\item{parMap}{a function that translates ABC variables (parABC)
into something the model will accept.}

\item{Phi}{Volume}

\item{parameters_from_expressions}{a vector of evaluated expressions}

\item{nStochSim}{number of stochastic simulations to average over}

\item{reactions}{a list that encodes the reactions for
GillespieSSA2}

\item{modelName}{model name}

\item{distance}{a user supplied function that calculates a distance
between simulation and data with an interface of
distance(simulation, data, errVal), where errVal is an estimate
of the measuremnet noise (e.g. standard deviation), if needed
by the function.}
}
\value{
the distance between the trajectory just simulated and the experiment considered
}
\description{
Function that simulates a stochastic trajectory with the Gillespie algorithm
given certain experimental conditions and a parameter vector, and computes the
distance between the simulation and the experimental data
}
