% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{simulator.stoch}
\alias{simulator.stoch}
\title{Function that creates a closure that simulates a stochastic trajectory with the Gillespie algorithm
given certain experimental conditions and a parameter vector, and computes the
distance between the simulation and the experimental data}
\usage{
simulator.stoch(
  experiments,
  model.tab = model.tab,
  reactions = NULL,
  parMap = identity,
  outputFunction = function(t, state, param) {
     state
 },
  vol = 4e-16,
  unit = 1e-06,
  nStochSim = 3,
  distance = NULL
)
}
\arguments{
\item{model.tab}{an SBtab model}

\item{reactions}{a list that encodes the reactions for
GillespieSSA2}

\item{parMap}{a function that translates ABC variables (parABC)
into something the model will accept.}

\item{vol}{Volume in which the reactions take place}

\item{unit}{unit of measure (1e-6 corresponds to micrometer)}

\item{nStochSim}{number of stochastic simulations to average over}

\item{experiment}{an experiment}
}
\value{
a function that given a parameter returns a simulated trajectory obrained via the Gillespie algorithm
}
\description{
Function that creates a closure that simulates a stochastic trajectory with the Gillespie algorithm
given certain experimental conditions and a parameter vector, and computes the
distance between the simulation and the experimental data
}
