% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{small}
\alias{small}
\title{This function reduces the sample to its effective size}
\usage{
small(
  S,
  L = attr(S, "logLikelihood"),
  verbose = getOption("uqsa.verbose", default = interactive())
)
}
\arguments{
\item{S}{an MCMC sample}

\item{L}{the log-likelihood values of S}

\item{verbose}{when TRUE the acf plot option is set to TRUE, and
teh found auto-correlation length is printed.}
}
\value{
a smaller version of S
}
\description{
When plotting, we want to show only a few representative
lines or points derived from a sample. This function will determine
the auto-correlation length very roughly and use that number to
thin out the sample to a minimal size that still represents the
original sample well.
}
\examples{
S <- matrix(rnorm(300),100,3)
## the next line fakes auto-correlation:
attr(S,"logLikelihood") <- cos(seq(0,1,length.out=100)) + rnorm(100,sd=0.05)
print(dim(S))
print(dim(small(S)))
}
