% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{so.path<-}
\alias{so.path<-}
\title{Add information about compiled code}
\usage{
so.path(o) <- value
}
\arguments{
\item{o}{the ode (list of named arrays and matrices)}

\item{value}{the path to the compiled model}
}
\value{
modified o, with information about compiled code
}
\description{
Adds the path of the shared library (.so file) to the ODE model.
}
