% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{so_path<-}
\alias{so_path<-}
\title{Add information about compiled code}
\usage{
so_path(o) <- value
}
\arguments{
\item{o}{the ODE (list of named arrays and matrices), or CME model}

\item{value}{the path to the compiled model}
}
\value{
modified o, with information about compiled code
}
\description{
Adds the path of the shared library (.so file) to the ODE model.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
c_path(o) <- write_c_file(generate_code(o))
so_path(o) <- shlib(o)
print(o)
}
