% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{stoichiometric_matrix}
\alias{stoichiometric_matrix}
\title{The stoichiometric matrix of a reaction network}
\usage{
stoichiometric_matrix(m, compound.names = rownames(m$Compound))
}
\arguments{
\item{m}{list of data frames with at least the 'Reaction' table,
and the 'Compound' table}

\item{compound.names}{all names of the reacting compounds}
}
\value{
the stoichiometric matrix, with some additional attributes.
}
\description{
Given a model, described in tabular form (\code{m} is a list of
data-frames). The stoichiometric matrix is the linear map between
the model's flux vector and the ODE's right-gand-side vector field.
If the flux vector is \code{rr <- flux(t,x,p)}, which maps the state
variables \code{x} and parameters \code{p} to the reaction rate \code{rr} of each
reaction. The stoichiometric matrix \code{nu} (\eqn{\nu}{ν}), will map the reaction
rates to the rate of change of the state variables: dx/dt := nu \%*\%
flux(t,x,p).
}
\details{
The matrix is usually sparse, but not extremely big. This function
attaches a sparse version of the same information as attributes to
the return-value, as two lists, for convenience.
}
\examples{
the_reaction <- "A + B <=> C"
m <- list(
    Reaction=data.frame(reactants=c("A+B"),products=c("C"))
)
nu <- stoichiometric_matrix(m,c("A","B","C"))
}
