% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{[.experiment}
\alias{[.experiment}
\title{Subset experiments with preserved class}
\usage{
\method{[}{experiment}(x, i, ...)
}
\arguments{
\item{x}{an object with class "experiment"}

\item{i}{an index-set}

\item{...}{passed on the list-[ function}
}
\value{
subset of experiments
}
\description{
The normal list subset operation would drop the class from the
experiment object (which is fine in theory). With this override,
the class is preserved.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
x <- experiments(m)
class(x)
class(x[seq(2)])
}
