% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{[.simulation}
\alias{[.simulation}
\title{Subset simulations with preserved class}
\usage{
\method{[}{simulation}(x, i, ...)
}
\arguments{
\item{x}{an object with class "simulation"}

\item{i}{an index-set}

\item{...}{passed on the list-[ function}
}
\value{
subset of experiments
}
\description{
The normal list subset operation would drop the class from the
simulation object (which is fine in theory). With this override,
the class is preserved.
}
