% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sum.of.bin.variance}
\alias{sum.of.bin.variance}
\title{Weighted Sum of Bin-specific variances}
\usage{
sum.of.bin.variance(hst, binMeans, totalMean)
}
\arguments{
\item{hst}{the histogram of the parameter sample}

\item{binMeans}{the means of the observable within each bin (rows of means)}

\item{totalMean}{the mean of the observable over the entire sample (vector)}
}
\value{
The weighted sum of square differences between the binMean and the totalMean
}
\description{
This function calculates the variance sum of a vector valued
observable.
}
