% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{swap_points_locally}
\alias{swap_points_locally}
\title{Swap the end-points of two Markov chains}
\usage{
swap_points_locally(parMCMC)
}
\arguments{
\item{parMCMC}{a list of Markov chain end points, each entry
annotated with a temperature attribute
\code{attr(parMCMC[[i]],"beta")}}
}
\value{
a list with some members swapped
}
\description{
This is a conditional swap, according to the rules of parallel
tempering. This function is only useful if the Markov chains have
returned to the global scope and one process will make the decision
and perform the swap, i.e.: the current state of each chain is
locally available.
}
