% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unit.info}
\alias{unit.info}
\title{Prints an interpretation string of a unit}
\usage{
unit.info(unit.str, unit = unit.from.string(unit.str))
}
\arguments{
\item{unit.str}{unit string}

\item{unit}{optionally, the data.frame that describes the unit}
}
\description{
given a string describing a unit of measurement, this function
prints the interpretation on screen, rather than returning it as a
data.frame
}
\examples{
> unit.info("km/h",unit.from.string("km/h"))
«km/h» has been interpreted as:
(1 × metre×10^(3))^(1)
(60 × second×10^(0))^(-1)
}
