% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{unit_as_character}
\alias{unit_as_character}
\title{converts a unit data.frame into a printable string}
\usage{
unit_as_character(unit)
}
\arguments{
\item{unit}{a data.frame created by unit.from.string()}
}
\value{
a string representation of that data.frame purely for printing
}
\description{
This is a crude function to make a printable representation of a
unit data.frame, with very explicit parentheses and exponents.
}
\examples{
u <- unit.from.string("s^-1")
str <- unit_as_character(u)
print(str)
}
