% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{units_from_table}
\alias{units_from_table}
\title{Get units from a data.frame column}
\usage{
units_from_table(df, default = "1")
}
\arguments{
\item{df}{a data.frame}

\item{default}{default value if no unit column exists}
}
\value{
a character vector of units with names
}
\description{
Given a data.frame this function retrieves the strings in the unit
column named: unit, Unit, units (partial matching disregarding
capitalization).
}
\details{
The returned value uses the row names of the data.frame as names of
the character vector of units.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAP79"))
u <- units_from_table(m$Compound)
}
