% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{update_values}
\alias{update_values}
\title{Updates the named values of vector \code{v} with values mentioned in data.frame d}
\usage{
update_values(v, d, as_type = "numeric")
}
\arguments{
\item{v}{a named vector}

\item{d}{data.frame with column names that correspond to those of \code{v}}

\item{as_type}{a character scalar indicating a type ('character','numeric','logical',etc.)}
}
\value{
a matrix of dimension length(v) × NROW(d)
}
\description{
Given a named vector, this function will create a mtarix, with
several copies of the vector, with different values, dependeing on
some context. The context is each row of a \code{data.frame}, with
columns referring to the names of \code{v}, and values for the entries
in \code{v} for that row's context. Example: the rows can be different
experinemnts, with each experiment assigning new values to some
members of \code{v} (but not necessarily all).
}
\examples{
\dontrun{
f <- uqsa_example("AKAR4")
m <- model_from_tsv(f)
iv <- values(m$Compound)
IV <- update_values(iv,m$Experiments)
print(IV)
}
}
