% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqsa_example_data.R
\name{uqsa_example}
\alias{uqsa_example}
\title{Load an example model for this package}
\usage{
uqsa_example(modelName = NULL, full.names = TRUE, pat = "[.]tsv$")
}
\arguments{
\item{modelName}{name of model, e.g.: "AKAR4", "AKAP79", "CaMKII";
if empty, this function lists all available examples.}
}
\value{
The location of the example's directory (where the model is
    stored).
}
\description{
This function finds the path to an example model, given by name.
In the SBtab format, model and data travel together (in different
tables, but the same documents).
}
