% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{values}
\alias{values}
\title{Find values in a data.frame that is derived from a tsv file or similar}
\usage{
values(df)
}
\arguments{
\item{df}{a data frame with a "value" column}
}
\value{
a named numeric vector
}
\description{
given a data.frame, this function will look for a column that
contains some kind of value and retrieve it. The returned numeric
vector will be named.
}
\details{
If the values contain a standard error, the returned value is of
class \emph{errors}.
}
