% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{write_and_compile}
\alias{write_and_compile}
\title{Writes code to file and compiles}
\usage{
write_and_compile(C, modelname = comment(C))
}
\arguments{
\item{C}{character vector with the code of this model}
}
\value{
the model's name with annotation about file names.
}
\description{
This function accepts the code that was written by \link{generate_code},
possibly changed by the user. It writes the contents to a c file
named 'modelName_gvf.c'. This file is compiled to './modelName.so'
using normal command line tools, not \verb{R CMD SHLIB}
}
\details{
This entire function can be replaced with a call to \code{cat()} and
then compiling the written file in the system's shell.
}
\examples{
\dontrun{
  cCode <- generate_code(m,o)            # a character vector
  modelName <- write_and_compile(cCode)  # commented name
  print(comment(modelName))
}
}
