% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_c_code}
\alias{write_c_code}
\title{Write the C code to a file}
\usage{
write_c_code(
  C,
  model.name = comment(C),
  file = file.path(tempdir(), digest::digest(C, "xxhash64"), paste0(model.name, ".c"))
)
}
\arguments{
\item{C}{the code to write, as a character array.}

\item{model.name}{a string with no special characters, will be used in the file name}

\item{file}{override the default file name (based on hashing)}
}
\value{
the path of the written file
}
\description{
This function does not compile the code, it only writes it to a
file in a temporary location (tempdir). By default, the name of the
file will contain the hash of the entire code.
}
\details{
If instead of a character vector, an ode or cme object is passed,
this function will generate code from it with default options.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
C <- generate_code(o)
c_path(o) <- write_c_code(C)
print(o)
if (file.exists(c_path(o))) cat("c file exists.\n")
}
