% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{writeCFunction}
\alias{writeCFunction}
\title{Write C Function}
\usage{
writeCFunction(
  prefix,
  fName,
  defArgs = c("double t", "const double y_[]"),
  retValue = NULL,
  defs = NULL,
  values = NULL,
  body = NULL,
  otherArgs = "void *par",
  init0 = TRUE
)
}
\arguments{
\item{prefix}{function's prefix (name of model)}

\item{retValue}{name of return value (attr: memset?)}

\item{defs}{optional expressions, parameters, etc.}

\item{values}{a character vecotor with values or mathematic formulae to return.}

\item{body}{a character vector of additional content for the body of the function}

\item{otherArgs}{additional arguments}

\item{fNname}{function's name}
}
\description{
This function writes a C function according to a typical template.
}
