% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{yJacobian}
\alias{yJacobian}
\title{Jacobian of string-math}
\usage{
yJacobian(f, x)
}
\arguments{
\item{f}{a character vector of length n}

\item{x}{a character vector of length m}
}
\value{
a character matrix (n×m) with derivatives df\link{i}/dx\link{j}
}
\description{
Given a named character array of math expressions and a vector of
independent variables, this function calculates the Jacobian matrix
of the math expressions with respect to the variables.
}
