% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCMCMCFunctions.R
\name{abc_mcmc}
\alias{abc_mcmc}
\title{Performs and Approximate Bayesian Computation Sampling of Model Parameters}
\usage{
abc_mcmc(
  objectiveFunction,
  startPar,
  N,
  burnIn = N,
  Sigma0 = cov(t(startPar)),
  dprior,
  batchSize = 100 * NROW(Sigma0),
  parAcceptable = function(p) {
     all(is.finite(p))
 }
)
}
\arguments{
\item{objectiveFunction}{function that, given a (vectorial)
parameter as input, simulates the model, and outputs the
distance between experimental data and data simulated from the
model with the parameter provided in input}

\item{startPar}{starting values for the parameter vector, can (and
should) be a matrix with n columns, where ech column is a valid
parameter vector}

\item{N}{requested number of batches to return, the sample will be
of size \code{batchSize*N}}

\item{burnIn}{number of batches where the transition kernel will be
adjusted to achieve an acceptance rate of below 10\%.}

\item{Sigma0}{multivariate normal covariance of Markov chain
transition kernel, defaults to the covariance of the initial
parameters. If startPar is one vector, this matrix must be
provided explicitly.}

\item{dprior}{a function that returns prior probability density
values.}

\item{parAcceptable}{a function that can reject a parameter vector
early based on user-requirements. Has to return a scalar
Boolean.}
}
\value{
a list containing a sample matrix and a vector of scores
(values of delta for each sample)
}
\description{
Given a set of simulation experiments (list), a model, parameter
boundaries, this function will draw a sample of parameters from the
posterior probability density of the given problem.
}
\details{
Normally, ABC would produce a highly auto-correlated sample,
wasting lots of disk-space; \code{batchSize} can be used to thin out the
sample, recording every batchSize-th point: with \code{batchSize=100},
we perform 100 updates to the Markov chain variable and then save
the state to the sample. Higher batchSize numbers improve the
apparent quality of the sample, but create more work per sampled
point.
}
\examples{
\dontrun{
  f <- uqsa_example("AKAR4")
  m <- model_from_tsv(f)
  o <- as_ode(m)
  ex <- experiments(m,o)
  C <- generate_code(o)
  c_path(o) <- write_c_code(C)
  so_path(o) <- shlib(o)
  s <- simulator.c(ex,o,parMap=log10ParMap)
  objFunc <- makeObjective(ex,s)
  startPar <- log10(values(m$Parameter))
  lowerBound <- startPar - m$Paramster$stdv # m$Parameter$min
  upperBound <- startPar + m$Paramster$stdv # m$Parameter$max
  abcSample <- ABCMCMC(
    objFunc,
    startPar=values(m$Parameter),
    100,
    cov(rprior(1000)),
    delta=1,
    dprior=dUniformPrior(lowerBound,upperBound),
    batchSize = 100
  )
}
}
