% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{as_cme}
\alias{as_cme}
\title{Interprets the provided model as a stochastic model}
\usage{
as_cme(m)
}
\arguments{
\item{m}{list of data.frames, obtained via \code{model_from_tsv()}}
}
\value{
a list containing the interpreted model.
}
\description{
The chemical master equation can be simulated as a Markov jump
process (or continuous time Markov chain). One of the stochastic
solver algorithms is the Gillespie algorithm. This function return
sa data structure that can be used to generate code for the
Gillespie solver in this package.
}
\details{
This function interprets the contionuous model \code{m} as a discrete
state model with molecule counts and propensities. For this reason,
we need to specify a volume for the simulations to take place in.

The model \code{m} is assumed to describe a reaction network, as a list
of data.frames (as retuned by \link{model_from_tsv}).  The systems
biology information in the file is assumed to be concentrations and
rate coefficients, regardless of the interpretation this function
will derive from it. This is to make the model format of the TSV
file fairly uniform and independent of how we want to solve the
derived equations, be it ODE or CME.

Like the ode object, the returned object can also store the paths
of files we create for this model, with: \link{c.path<-}, and \link{so.path<-}

With the information provided with the rate coefficient units and a
volume, this function tries to convert everything to Gillespie rate
constants.
}
