% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{c.path<-}
\alias{c.path<-}
\title{Add information about the model's C code}
\usage{
\method{c}{`path<-`}(o, value)
}
\arguments{
\item{o}{the ODE , or CME model}

\item{value}{the path to the compiled model}
}
\value{
modified o, with information about compiled code
}
\description{
Adds the location of the model's C code (a file). The model is
typically a list of named numeric and named character vectors,
which describe the (interpreted) model.
}
