% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{c_path<-}
\alias{c_path<-}
\title{Add information about the model's C code}
\usage{
c_path(o) <- value
}
\arguments{
\item{o}{the ODE , or CME model}

\item{value}{the path to the compiled model}
}
\value{
modified o, with information about compiled code
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
c_path(o) <- write_c_code(generate_code(o))
so_path(o) <- shlib(o)
print(o)
}
\description{
Adds the location of the model's C code (a file). The model is
typically a list of named numeric and named character vectors,
which describe the (interpreted) model.
}
