% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{change_temperature}
\alias{change_temperature}
\title{Should 2 Markov chains exchange their temperatures}
\usage{
change_temperature(b1, ll1, b2, ll2)
}
\arguments{
\item{b1}{the inverse temperature of chain 1}

\item{ll1}{the log-likelihood of chain 1}

\item{b2}{the inverse temperature of chain 2}

\item{ll2}{the log-likelihood of chain 2}
}
\value{
TRUE is the chains should swap their temperatures
}
\description{
This function makes a Boolean choice about changes in temperature,
based on the log(likelihood) values of two Markov chains in a
parallel tempering setting. The outcome is stochastic.
}
\details{
This function is useful if \code{mpi.send()} and \code{mpi.recv()} are used.
}
\examples{
b <- c(1.0,0.5)
if (change_temperature(b[1],-850,b[2],-600)){ # with some randomness
  message(sprintf("yes, swapping temperature \%f <=> \%f",b[1],b[2]))
} else {
  message(sprintf("no, temperature \%f, and \%f stay unchanged",b[1],b[2]))
}
}
