% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{conservation_law_analysis}
\alias{conservation_law_analysis}
\title{Reduce the size of the system}
\usage{
conservation_law_analysis(nu, iv, verbose = FALSE)
}
\arguments{
\item{nu}{stoichiometric matrix}

\item{iv}{initial values}

\item{verbose}{if \code{TRUE}, this function will print the conservation laws on screen}
}
\value{
a list of conservation laws
}
\description{
Given a stoichiometric matrix, this function performs model
reduction via linear algebra operations, with \link[pracma:null]{pracma::null}.
}
\examples{
f <- uqsa_example("AKAR4")
m <- model_from_tsv(f)
nu <- stoichiometric_matrix(m)
CL <- conservation_law_analysis(nu,values(m$Compound))
print(names(CL))
print(CL[,c('value','Formula')])
}
