% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{conservation_law_analysis}
\alias{conservation_law_analysis}
\title{Reduce the size of the system}
\usage{
conservation_law_analysis(nu, iv, verbose = FALSE)
}
\arguments{
\item{nu}{stoichiometric matrix}
}
\value{
a list of conservation laws
}
\description{
Given a stoichiometric matrix, this function performs model
reduction via linear algebra operations, with \link[pracma:nullspace]{pracma::null}.
}
