% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCRNN.R
\name{CRNN}
\alias{CRNN}
\title{CRNN creates C code for a chemical reaction neural network}
\usage{
CRNN(numReactions, initialValues, funcValues, model.name = "CRNN")
}
\arguments{
\item{numReactions}{the number of reversible mass action law
reactions}

\item{initialValues}{named vector of initial values, names will be
used as the neames of the reacting compounds.}

\item{funcValues}{named character vector, can be any valid C
expression (one line) of the available state variables (the
names can be used literally).}

\item{model.name}{the prefix of all created model functions: CRNN_vf, CRNN_jac, ...}
}
\value{
a character vector suitable for writing to a file (.c)
}
\description{
This function creates a very general ODE (c source code), that can
be compiled and simulated using the UQSA package.
}
\details{
Example: A + B <=> C
numReactions: n <- 1
initialValues: x <- c(A=2,B=3,C=0)
funcValues: f <- c("A+B","log(A)")

The above definition would create a CRNN inspired ODE, where A+B
and log(A) are treated as observable (measureable) values
(functions of the state variables).

Note: In addition to the state variables, the function values can
also reference the log-parameters of the model as \code{l[i]}, where \code{i} is
a 0-based offset to the reaction \code{i}; the backward rate, is stored at
position \code{l[i+numRct]}.
}
\examples{
C <- CRNN(4,c(A=1,B=2,C=3),c(out="A+B+C"),model.name="testmodel")
cat(head(C),sep='\n')
}
