% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{dCopulaPrior}
\alias{dCopulaPrior}
\title{copulaPrior creates a prior probability density function}
\usage{
dCopulaPrior(Copula)
}
\arguments{
\item{Copula}{a list, as returned by fitCopula() or makeIndepCopula}
}
\value{
a function that maps parameters (a vector) to probability density values (scalar)
}
\description{
This function accepts the return list of fitCopula() or
makeIndepCopula() and creates a density function from it.
}
\examples{
x<-rnorm(300,mean=1,sd=2)
X<-matrix(x,100,3)
C<-fitCopula(X)
d<-dCopulaPrior(C)
print(d(c(1,2,3)))
print(prod(sapply(c(1,2,3),FUN=dnorm,mean=1,sd=2)))
}
