% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{defaultDistance}
\alias{defaultDistance}
\title{default distance function for one experiment}
\usage{
defaultDistance(funcSim, dataVAL, dataERR = max(dataVAL))
}
\arguments{
\item{funcSim}{a matrix, contains model solution (output values),
columns of output vectors}

\item{dataVAL}{a matrix of experimental data, shaped like funcSim}

\item{dataERR}{a matrix of measurement errors, if available,
defaults to the maximum data value.}
}
\value{
a numeric scalar, the distance between data \code{dataVAL} and
simulation \code{funcSim}.
}
\description{
if each experiment corresponds to one simulation and is fully
quanitified by itself, then calculating the overall distance
between data and experiment can be done one by one. This function
describes the default way a simulation is compared to data.
}
\details{
If the data is more complex, and two or more simulations are needed
to calculate one distance value then the objective-Function needs
to be entirely user-supplied. This is the case with experiments
that have a "control" -- this is needed when the measurement is in
arbitrary units and only makes sense comparatively to a secondary
(control) scenario.

This function will be used if none is provided by the user.

The funcSim values need to be supplied as a matrix of size N×T with
N the length of the model's output vectors and T the amount of
measurement times (this is how the rgsl package returns the
simulation results).
}
\examples{
d <- defaultDistance(seq(7),seq(7)+rnorm(7,0,0.1),rep(0.1,7))
}
