% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{experiments}
\alias{experiments}
\title{Extract Measured Data and Simulation Experiment Instructions}
\usage{
experiments(m, o)
}
\arguments{
\item{m}{the model (with data), as obtained via \code{model_from_tsv()}, or similar.}

\item{o}{the ode derived from \code{m}}
}
\value{
a list of simulation instructions
}
\description{
This function accepts the model obtained via \code{model_fromt_tsv} or a
similar function. It finds the data tables for this model (if any
are present), and finds the simulation instructions to reproduce
these data sets using the model.
}
\details{
This function requires that the files the model is stored as
contains measurements (data) that can be interpreted fairly
easily. Each data file needs columns that are named like the
observable quantities listed in the Output table.

If the data is very indirectly related to the model, then we don't
interpret the data files themselves and the user needs to write a
specialized likelihood function to relate the raw data in the files
with something that the model does. In such cases, don't use this function.

The instructions must be organised in a table called Experiment(s).
}
\examples{
f <- uqsa_example("AKAR4")
m <- model_from_tsv(f)
o <- as_ode(m)
ex <- experiments(m,o)
print(names(ex))
print(ex[[1]]$input)
}
