% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaFunctions.R
\name{fitCopula}
\alias{fitCopula}
\title{Makes a Probability Density Estimate (from a sample)}
\usage{
fitCopula(X)
}
\arguments{
\item{X}{sample that characterizes the traget distribution (rows)}
}
\value{
as list: vineCop, U, Z, and Y where U are marginal
probability samples, Z are cummulative density values for U,
and Y are the probability density values of U.
}
\description{
Given a sample (from some probability distribution) this function
makes a Copula fit to the source distribution using the VineCopula
package.
}
\examples{
rprior <- rNormalPrior(c(1,2,3),c(4,5,6))
X <- rprior(1000)
C <- fitCopula(X)
rCopula <- rCopulaPrior(C)
Z <- rCopula(1000)
print(norm(cov(X) - cov(Z),"2")/norm(X,"2"))
print(abs(sum(colMeans(X) - colMeans(Z)))/sum(colMeans(X)))
}
