% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulaFunctions.R
\name{fitCopula}
\alias{fitCopula}
\title{Makes a Probability Density Estimate (from a sample)}
\usage{
fitCopula(X, nCores = detectCores())
}
\arguments{
\item{X}{sample that characterizes the traget distribution (rows)}

\item{nCores}{passed to parallel::mclapply()}
}
\value{
as list: vineCop, U, Z, and Y where U are marginal
    probability samples, Z are cummulative density values for U,
    and Y are the probability density values of U.
}
\description{
Given a sample (from some probability distribution) this function
makes a Copula fit to the source distribution using the VineCopula
package.
}
