% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{ftsplit}
\alias{ftsplit}
\title{Fixed-token-split a scalar string into parts}
\usage{
ftsplit(str, s)
}
\arguments{
\item{str}{a string (character vector of length 1)}

\item{s}{a split token}
}
\value{
a character vector of the components without leading or trailing whitespace
}
\description{
This function splits a string like strsplit, but it removes
whitespace from the result (on both ends) and the split token is
taken literally
}
\examples{
ftsplit(" A + 2*B ","+")
[1] "A"   "2*B"
x<-c('a+b','c+d','1 + 2'); lapply(x,ftsplit,'+')
[[1]]
[1] "a" "b"
[[2]]
[1] "c" "d"
[[3]]
[1] "1" "2"
this also works, but mixes up the components:
x<-c('a+b','c+d+1','1 / 2'); ftsplit(x,'+')
[1] "a"     "b"     "c"     "d"     "1"     "1 / 2"
}
