% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{gatherReplicas}
\alias{gatherReplicas}
\title{gatherReplicas collects all sample points, from all files, which
are assumed to be exact replicas, with different seeds (and
possibly sizes). This function uses mclapply to process the files,
which may be quicker than gatherSample.  The temperature is
disregarded, assuming that no parallel tempering was used.  To
facilitate the loading of a very big sample, this function will
analyse the auto-correltation within each file and returned a
thinned subsample of size N/2*tauint (effective sample size). There
is no need to further reduce the rfesult.}
\usage{
gatherReplicas(files)
}
\arguments{
\item{files}{a list of file names}
}
\value{
a matrix of sampled points, all with the same temperature
}
\description{
For small samples, it is better to load the entire sample and
analyse it in full. This function is intended for samples that are
so big that they challenge the memory of the machine.
}
\details{
This function is quicker if you have used trivial parallelism,
without mpi communication between the ranks (or another method of
obtaining several replicas, like forking or sequetial reprtition).

This function assumes that each supplied RDS file contains a matrix
of model MCMC parameters. The returned value X will be the rbind of
all the smaller x contained in the individual files. The value X
will have several attributes attached to it:
\itemize{
\item logLikelihood: log(likelihood(X\link{i,})), one value per row of X
\item stepSize: the MCMC step size used in each given file#'
}
}
