% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{gatherSample}
\alias{gatherSample}
\title{gatherSample collects all sample points, from all files, with the
given temperature}
\usage{
gatherSample(files, beta = 1, size = NA)
}
\arguments{
\item{files}{a list of file names}

\item{beta}{the inverse temperture to extract sample for}

\item{size}{a size the is smaller than the actual sample size, if
left unchanged, all sampled points are returned}
}
\value{
a matrix of sampled points, all with the same temperature
}
\description{
This function assumes that each supplied RDS file contains a matrix
of model MCMC parameters, with an attribute called "beta" that
lists the temperature of each row.
}
\details{
This function selects and collects all rows, from all files with
the same (given) temperature.
}
