% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{generateCode}
\alias{generateCode}
\title{Write C code}
\usage{
generateCode(ode)
}
\arguments{
\item{ode}{a list of data.frames with the text representation of the ordinary differential equation}
}
\value{
a character vector with the generated code, one element is one line.
}
\description{
Write C code
}
