% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{gllf}
\alias{gllf}
\title{Default gradient-log-likelihood Function}
\usage{
gllf(parMapJac = log10ParMapJac)
}
\arguments{
\item{parMCMC}{a numeric vector, with attributes for MCMC, specifically smmala}
}
\value{
a numeric vector: grad(log(likelihood(data|parMCMC)))
}
\description{
Extracts the \code{FisherInformation} values from the simulations attribute
of the parMCMC argument, requires:
\itemize{
\item parMCMC has simulations attribute
\item simulations list includes Fisher-Information values
}
}
\details{
This function will take the log-likelihood gradient values
claculated by the ode solver in this package, and return the sum of
those vectors over all experiments. The gll-value the simulator
returns is calculated with the assumption of a normal distribution
on measurement errors.

This function uses the log10ParMapJac(parMCMC) function by default,
which assumes that sampling takes place in logarithmic space with
base 10.

Like \link{ll} this function does almost no work, it merely sums up the
gradient values calculated during simulation, but it also performs a
transformation of the gradient vector, taking the
parameter-mapping between the sampling-space and
model-parameter-space into account.
}
