% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{gllf}
\alias{gllf}
\title{Default Fisher Information Function}
\usage{
gllf(parMapJac = log10ParMapJac)
}
\arguments{
\item{parMCMC}{a numeric vector, with attributes for MCMC, specifically smmala}
}
\value{
a numeric vector: grad(log(likelihood(data|parMCMC)))
}
\description{
Extracts the \code{FisherInformation} values from the simulations attribute
of the parMCMC argument, requires:
\itemize{
\item parMCMC has simulations attribute
\item simulations list includes Fisher-Information values
}
}
\details{
This function will take the fi-values claculated by the ode solver
in this package, and return the sum of those values over all
experiments. The gll-value the simulator returns is calculated with
the assumption of a normal distribution on measurement errors.

This function uses the log10ParMapJac(parMCMC) function by default,
which assumes that sampling takes place in logarothmic space.
}
