% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{\%as\%}
\alias{\%as\%}
\title{\%as\% is a binary operator on strings with units in them}
\usage{
txtUnit \%as\% target
}
\arguments{
\item{txtUnit}{a string with numeric values, including units,
e.g. "3 cm"}

\item{target}{string, target unit, e.g. "m", must be scalar}
}
\value{
a numeric value y: val\emph{originalUnit = y}targetUnit, the
target unit is attached to the returned value, as a comment.
}
\description{
The function calls the units utility and converts the string on the
left into the unit on the right, e.g.: "cm" \%as\% "inches", both
units can contain numbers. Any input that is accepted by the units
utility is acceptable, as long as it makes sense with the command
line arguments: \verb{units -t "$originalUnit" "$targetUnit"}
}
\details{
FreeBSD and MacOS have a buggy \code{units} command: it fails to signal
failure, so this function does some additional checks to make sure
that the conversion worked.
}
\examples{
  ## needs `units` utility (a system utility)
  if (nzchar(Sys.which("units"))){
      y <- "21 cm" \%as\% "inches"
      print(y)
      y <- "12 nmol/L" \%as\% "mol/L"
      print(y)
      y <- "12 mol/m^3" \%as\% "mmol/L"
      print(y)
  }
}
