% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{\%@\%}
\alias{\%@\%}
\title{Fetch an Attribute}
\usage{
x \%@\% a
}
\arguments{
\item{x}{an R object (variable with attributes)}

\item{a}{the name of an attribute}
}
\value{
the value of the attribute: \code{attr(x,a)}
}
\description{
This function differs from \verb{rlang::\%@\%} in that it stops if the
attribute doesn't exist.
}
