% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{\%has\%}
\alias{\%has\%}
\title{checks whether a variable has the named attributes}
\usage{
var \%has\% attrNames
}
\arguments{
\item{var}{a variable to check for attributes}

\item{attrNames}{named attributes}
}
\value{
TRUE if all attributes are present
}
\description{
checks whether a variable has the named attributes
}
