% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{gsa_binning}
\alias{gsa_binning}
\title{Global Sensitivity Analysis}
\usage{
gsa_binning(parSample, outputSample, nBins = "Sturges")
}
\arguments{
\item{parSample}{a matrix of parameter vectors (rows)}

\item{outputSample}{a matrix, with rows of outputs (row-index is the sample index)}

\item{nBins}{number of bins, if unset defaults to the default of the hist function}
}
\value{
sensitivity \code{S[i,j]} of \code{output[i]} with respect to \code{parameter[j]}
}
\description{
This function performs a binning based estimation of the global
sensitivity of a model's output with respect to the model's
parameters. The output can be a prediction of the model's behaviour
in a scenario of interest (parameters, input, intial values,
boundary conditions, scheduled events etc.). The output models a
potentially measurable value (the "observable"). The sample-rows
and the output rows must correspond (they must be from the same
model simulation).
}
\examples{
  rprior <- rNormalPrior(c(-1,0,1),c(1,2,3))
  X <- rprior(10000)
  colnames(X) <- LETTERS[seq(3)]
  Z <- exp(X[,1,drop=FALSE]+X[,2,drop=FALSE])
  colnames(Z) <- "alpha"
  GSA <- gsa_binning(X,Z)
  print(GSA)
  cat("global sensitivity of alpha with respect to B: ",GSA['alpha','B'],"\n")
}
