% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{gsl_odeiv2_CRNN}
\alias{gsl_odeiv2_CRNN}
\title{simulates a CRNN ode model with extra work}
\usage{
gsl_odeiv2_CRNN(
  name,
  experiments,
  l,
  nu,
  m,
  abs.tol = 1e-06,
  rel.tol = 1e-05,
  initial.step.size = 0.001,
  method = 0
)
}
\arguments{
\item{name}{the name of the ODE model to simulate (a shared library
of the same name will be dynamically loaded and needs to be
created first)}

\item{experiments}{a list of \code{N} simulation experiments (time,
parameters, initial value, events).}

\item{l}{a matrix of parameters with M columns, in log-space.}

\item{nu}{a stoichiometry matrix (N×R) where N is the number of state
variables and R the number of reactions, all reactions are
assumed to be reversible.}

\item{m}{modifiers -- similar to stoichiometry, but indicates whether the
species takes part in the reaction without being consumed.}

\item{abs.tol}{absolute tolerance, real scalar.}

\item{rel.tol}{relative tolerance, real scalar.}

\item{initial.step.size}{initial value for the step size; the step
size will adapt to a value that observes the tolerances, real
scalar.}
}
\value{
a list of the solution trajectories \verb{y(t;p)} for all
experiments (named like the experiments), as well as the output
functions.
}
\description{
This function calls a C function which solves an initial value
problem, derived from a CRNN.
}
\keyword{ODE}
