% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{gsl_odeiv2_fi}
\alias{gsl_odeiv2_fi}
\title{simulates an ode model with extra work}
\usage{
gsl_odeiv2_fi(
  name,
  experiments,
  p,
  abs.tol = 1e-06,
  rel.tol = 1e-05,
  initial.step.size = 0.001,
  method = 0,
  omit = 0
)
}
\arguments{
\item{name}{the name of the ODE model to simulate (a shared library of the same name will be dynamically loaded and needs to be created first)}

\item{experiments}{a list of \code{N} simulation experiments (time, parameters, initial value, events)}

\item{p}{a matrix of parameters with M columns}

\item{abs.tol}{absolute tolerance, real scalar}

\item{rel.tol}{relative tolerance, real scalar}

\item{initial.step.size}{initial value for the step size; the step size will adapt to a value that observes the tolerances, real scalar}

\item{omit}{an integer that indicates how many of these to omit in this order: fisher information, gradient of the log-likelihood, log-likelihood}
}
\value{
a list of the solution trajectories \verb{y(t;p)} for all experiments (named like the experiments), as well as the output functions
}
\description{
This function calls a C function which solves an initial value
problem, calculates the sensitivity of the solution, log-likelihood
value \code{ll}, gradient of \code{ll} amd Fisher-Information.
}
\keyword{ODE}
