% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{high_level_metropolis}
\alias{high_level_metropolis}
\title{High Level Metropolis function}
\usage{
high_level_metropolis(
  m,
  o = as_ode(m, cla = FALSE),
  ex = experiments(m, o),
  x = values(m$Parameter),
  beta = 1
)
}
\arguments{
\item{m}{the model's TSV representation read via \code{model_from_tsv}}

\item{o}{(optional) ode representation of \code{m}}

\item{ex}{experiments of \code{m}, with simulation instructions for \code{o}.}

\item{x}{initial point of the markov chain, pre in itialized to
have the right attributes.}

\item{beta}{for parallel tempering, the log-likelihood will have a
factor of \code{beta} applied to it}
}
\value{
\code{smmala} a function of three arguments: p0, N, eps; where
p0 is the starting point, N is the desired sample-size, and eps
is the step size. This function has an attribute called "init",
with a pre-initialized starting point.
}
\description{
This function uses default assumption everywhere and returns a
function that will sample from the given model. This funciton will
generate code, compile the code, create an ODE solver for it, infer
the sampling space from the scale of the parameters, create all
necessary functions to move in parameter space (gradients of
likelihood and prior), as well as Fisher Information functions.
}
\examples{
\donttest{
  m <- model_from_tsv(uqsa_example("AKAP79"))
  rwm <- high_level_metropolis(m) # "random walk", metropolis algorithm
  p <- rwm \%@\% "init"             # a valid starting point
  N <- 100
  smallSample <- rwm(rwm \%@\% "init",N,1e-6)
  plot(
    smallSample \%@\% "logLikelihood",
    type="l",
    main=sprintf("\%i iterations",N),
    xlab="iterations",
    ylab="log-likelihood"
  )
}
}
