% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{high_level_smmala}
\alias{high_level_smmala}
\title{High Level SMMALA function}
\usage{
high_level_smmala(m, o = as_ode(m, cla = TRUE), ex = experiments(m, o))
}
\arguments{
\item{m}{the model's TSV representation read via \code{model_from_tsv}}

\item{o}{(optional) ode representation of \code{m}}

\item{ex}{experiments of \code{m}, with simulation instructions for \code{o}.}
}
\value{
\code{smmala} a function of three arguments: p0, N, eps; where p0 is the starting point, N is the desired sample-size, and eps is the step size.
}
\description{
This function uses default assumption everywhere and returns a
function that will sample from the given model. This funciton will
generate code, compile the code, create an ODE solver for it, infer
the sampling space from the scale of the parameters, create all
necessary functions to move in parameter space (gradients of
likelihood and prior), as well as Fisher Information functions.
}
