% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{importReactionsSSA}
\alias{importReactionsSSA}
\title{Functions to construct and run the stochastic simulation using GillespieSSA2 package}
\usage{
importReactionsSSA(model.tab, compile = TRUE)
}
\arguments{
\item{compile}{boolean value: if TRUE (default value), the GillespieSSA2 reactions are compiled (recommended for faster simulations)}

\item{model}{the model, represented by a list of data.frames with SBtab content}
}
\value{
a list of GillespieSSA2::reaction items (if compile==TRUE, the reactions are compiled)
}
\description{
This translates the Reaction network into the specific form required by GillespieSSA2
}
