% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLD.R
\name{KLD}
\alias{KLD}
\title{Kullback Leibler Divergence}
\usage{
KLD(X, Y, de = c("copula", "ks", "mvtnorm"))
}
\arguments{
\item{X}{sample from distribution P}

\item{Y}{sample from distribution Q}

\item{de}{density estimation mechanism (character scalar)}
}
\value{
D, a scalar value, the Kullback Leibler Divergence
}
\description{
This function calculates the Kullback Leibler Divergence $D(P\|Q)$
value between two distributions $P$ and $Q$, represented by their
two samples, \code{X} and \code{Y}. Both samples will have their density
inferred. The intended use-case is to compare 2D and 3D densities,
e.g.: to find interesting pairs of parameters within a bigger
distribution.
}
\details{
This estimate requires a method of density estimation, by default
we use the copula based methods \link{fitCopula} and \link{dCopulaPrior}
(which has a fairly high accuracy, but can be quite slow).

Effects of setting \code{de} to
\itemize{
\item \code{"ks"}: density is estimated via \code{ks::kde()}
\item \code{"mvtnorm"}: density is estimated via manual kernel
density estimation, using the multivariate
Gaussian density \code{mvtnorm::dmvnorm}
}
}
