% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{linear_scale}
\alias{linear_scale}
\title{Interprets a character vector as names of logarithms}
\usage{
linear_scale(x, str_scale = attr(x, "scale"))
}
\arguments{
\item{x}{values}

\item{str_scale}{character vector}
}
\value{
a copy of \code{x},  transformed into linear space
}
\description{
The values in \code{x} are possibly given in a logarithmic space. The
parameter \code{str_scale} gives this logarithmic scale (provided in a
language agnostic form), by a human. An empty string causes no
transformations. Similarly, providing no scale at all causes no
transformations.
}
\details{
The words in \code{str_scale} name a logarithm, e.g. "log10". Currently understood scales:
\itemize{
\item log10
\item log2, ld
\item ln, log
}
}
\examples{
x <- c(1,2,3,1,1,1)
attr(x,"scale") <- c("log10","log2","log","ln","ld","log5")
print(linear_scale(x))
}
