% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{ll}
\alias{ll}
\title{Default Log-likelihood Function}
\usage{
ll(parMCMC)
}
\arguments{
\item{parMCMC}{a numeric vector, with attributes for MCMC, specifically smmala}
}
\value{
a scalar value: log(likelihood(data|parMCMC))
}
\description{
Extracts the \code{logLikelihood} value from the simulations attribute
of the parMCMC argument, requires:
\itemize{
\item parMCMC has simulations attribute
\item simulations list includes logLikelihood values
}
}
\details{
This function will take the log-likelihood-values claculated by the
ode solver in this package, and return the sum of those values over
all experiments. The value the simulator returns is calculated
with the assumption of a normal distribution on measurement errors.

This function does almost no work, it merely sums up the values
calculated during simulation.
}
