% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{loadSample_mpi}
\alias{loadSample_mpi}
\title{This function merges mpi-samples into one}
\usage{
loadSample_mpi(files)
}
\arguments{
\item{files}{the files where the individual samples are stored}
}
\value{
one matrix where all samples are concatenated.
}
\description{
When using MPI, we save the sample immediately into a file, each
rank saves to its own file. This function collects all of these
smaller samples into one. The samples should be saved with
\code{saveRDS()}.
}
