% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{log10ParMapJac}
\alias{log10ParMapJac}
\title{LOG10 parameter mapping, jacobian}
\usage{
log10ParMapJac(parMCMC)
}
\arguments{
\item{parMCMC}{the sampling variables (numeric vector)}
}
\description{
This map is used by the simulator to transform sampling variables
into ODE-model porameters. As we often calculate sensitivites, we
alos need the jacobian of the map, due to the chain rule of
differentiation.
}
