% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{log2ParMap}
\alias{log2ParMap}
\title{LOG2 parameter mapping used by the MCMC module}
\usage{
log2ParMap(parMCMC)
}
\arguments{
\item{parMCMC}{the sampling variables (numeric vector)}
}
\description{
This map is used by the simulator to transform sampling variables
into ODE-model porameters.
}
