% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{makeGillespieModel}
\alias{makeGillespieModel}
\title{makeGillespieModel interprets the provided SBtab file as a stochastic model}
\usage{
makeGillespieModel(sb)
}
\arguments{
\item{sb}{list of data.frames}

\item{LV}{Avogadro's constant L multiplied by the system's volume V.}
}
\value{
a list containing the interpreted model.
}
\description{
The SBtab file is assumed to describe a reaction network.  The
systems biology information is assumed to be concentrations and
rate coefficients.
}
\details{
With the information provided with the rate coefficient units and a
volume, this function tries to convert everything to Gillespie rate
constants.
}
