% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{makeGillespieModel}
\alias{makeGillespieModel}
\title{Create a list of reactions for GillespieSSA2}
\usage{
makeGillespieModel(SBtab)
}
\arguments{
\item{SBtab}{a series of tables as returned by `sbtab_from_tsv()`}
}
\value{
a list of reactions
}
\description{
This function takes a series of SBtab tables, as returned by
SBtabVFGEN::sbtab_from_tsv() and creates GillespieSSA2 reactions
from them. Reactions arfe made pairwise, as forward and backward
reaction pairs. If a backward reaction doesn't exist, the list item
is NULL. A valid set of reactions can be obtained with `!is.null(reactions)`
}
\examples{
 model.tsv<-dir(pattern="[.]tsv$")
 model.sbtab<-SBtabVFGEN::sbtab_from_tsv(model.tsv)
 reactions<-makeGillespieModel(model.sbtab)
 l<-is.null(reactions)
 model.ssa2<-reactions[!l]
}
