% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{makeObjective}
\alias{makeObjective}
\title{creates Objective functions from ingredients}
\usage{
makeObjective(experiments, simulate, distance = defaultDistance)
}
\arguments{
\item{experiments}{a list of simulation experiments}

\item{simulate}{closure that simulates the model}

\item{distance}{a function that calculates ABC scores (distance between data and simulations)}
}
\value{
an objective function
}
\description{
the returned objective function has only one argument: the ABC
variables that shall be mapped to ODE-model parameters.
}
\details{
The user supplied distance function should accept three arguments:
distance(SIM, DATA, STDV), all three matrices.  SIM is the model
output (simulation), DATA is the measured data, while STDV
represents the standard error of that measurement. All three have
the same size: N×M, where N is the number of observables (outputs),
and M is th enumber of measurement time-points (length of the
time-series).
}
