% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{makeObjective}
\alias{makeObjective}
\title{creates Objective functions from ingredients}
\usage{
makeObjective(
  experiments,
  modelName = NULL,
  distance,
  parMap = identity,
  simulate = NULL
)
}
\arguments{
\item{experiments}{a list of simulation experiments}

\item{modelName}{and model storage file as comment}

\item{distance}{a function that calculates ABC scores}

\item{parMap}{a function that transforms ABC variables into acceptable model parameters}

\item{simulate}{closure that simulates the model}
}
\value{
an objective function
}
\description{
the returned objective function has only one argument: the ABC
variables that shall be mapped to ODE-model parameters.
}
