% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{makeObjectiveSSA}
\alias{makeObjectiveSSA}
\title{Function that creates the objective function}
\usage{
makeObjectiveSSA(
  experiments,
  model,
  parNames,
  distance,
  parMap = identity,
  Phi,
  reactions,
  nStochSim = 1,
  parameters_from_expressions = NULL
)
}
\arguments{
\item{experiments}{a list of experiments}

\item{parNames}{the names of the (biological) parameters of the
model}

\item{distance}{a user supplied function that calculates a distance
between simulation and data with an interface of
distance(simulation, data, errVal), where errVal is an estimate
of the measuremnet noise (e.g. standard deviation), if needed
by the function.}

\item{parMap}{a function that translates ABC variables (parABC)
into something the model will accept.}

\item{Phi}{Volume}

\item{reactions}{a list that encodes the reactions for
GillespieSSA2}

\item{nStochSim}{number of stochastic simulations to average over}
}
\value{
a closure for the objective function that implicitly
depends on all of the arguments to this function but explicitly
only on the ABC parameters parABC.
}
\description{
Given a parameter set, this function computes the distance between
experimental data and simulated data (coresponding to the parameter
in input).
}
