% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Markov Chain Monte Carlo}
\usage{
mcmc(update, backingpath = NULL, backingfile = NULL, descriptorfile = NULL)
}
\arguments{
\item{update}{and update function}

\item{backingpath}{if this is set, a big.matrix will be used to
store the sample.}

\item{backingfile}{the name of the backing file for the big.matrix}

\item{descriptorfile}{the descriptor file-name for the big.matrix
object}
}
\value{
M(initiPar,N), a function (closure) of initial starting values and
number of Markov chain steps
}
\description{
This function creates an MCMC function for a given set of
experiments. The Markov chains have no communication between them
if more than one is created using this mechanism.
}
\details{
The algorithm is entirely determined by the update function.  Any
intermediate values that updates requires aside from simulation
results have to be attributes of the MCMC variable: parMCMC.

The update function: update(parGiven) -> parUpdate depends only on
the given parameters, all other dependencies have to be either
implicit (as a closure) or attributes of parGiven.

When backingpath and backingfile are both unset, a normal matrix is
returned. If these options are set, then the returned closure will
create \code{big.matrix} objects. The backingpath alone is sufficient to
create a \code{big.matrix}, with default names for the sample files.

The returned closure will itself either return (A) a matrix (a
sample) with N rows (one row per sample member), with attributes
related to how the Markov chain performed, or (B) a list with a
\link[bigmemory:big.matrix]{bigmemory::big.matrix} object for the sample and a second matrix with just
the MCMC attributes (such as log-likelihood values).
}
\examples{
\dontrun{
m <- model_from_tsv(uqsa_example("AKAP79"))
rwm <- high_level_metropolis(m) # "random walk", metropolis algorithm
p <- rwm \%@\% "init"             # a valid starting point
smallSample <- rwm(rwm \%@\% "init",600,1e-4)
pairs(smallSample[,seq(6)])
}
}
