% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_init}
\alias{mcmc_init}
\title{Initialize the Markov chain}
\usage{
mcmc_init(
  beta,
  parMCMC,
  simulate,
  logLikelihood = ll,
  dprior = function(x) prod(rnorm(x)),
  gradLogLikelihood = NULL,
  gprior = NULL,
  fisherInformation = NULL
)
}
\arguments{
\item{beta}{inverse temperature for the Markov chain (parallel
tempering)}

\item{parMCMC}{a plain starting value for the Markov chain}

\item{simulate}{a closure that maps the MCMC variable to simulation
results (the simulation experiments are enclosed in this
function).}

\item{logLikelihood}{a function that maps simulations to
logLikelihood values}

\item{dprior}{density of the prior distribution}

\item{gradLogLikelihood}{the gradient function of the logLikelihood
(optional) -- only if the algorithm requires it}

\item{gprior}{the gradient pf the log-prior (for SMMALA and similar
algorithms).}

\item{fisherInformation}{a function that calculates the Fisher
Information matrix}
}
\value{
the same starting parameter vector, but with attributes.
}
\description{
This function must append all required attributes to the MCMC
varible, for the Markov chain to update correctly.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
p0 <- values(m$Parameter)
c_path(o) <- write_c_code(generate_code(o))
so_path(o) <- shlib(o)
ex <- experiments(m,o)
s <- simfi(ex,o)
dprior <- dNormalPrior(p0,m$Parameter$stdv)
p <- mcmc_init(1.0,p0,s,dprior=dprior)
print(names(attributes(p))) ## now has attributes necessary for MCMC
}
