% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc_mpi}
\alias{mcmc_mpi}
\title{The MPI version of the mcmc function}
\usage{
mcmc_mpi(
  update,
  comm,
  swapDelay = 0,
  swapFunc = pbdMPI_bcast_reduce_temperatures
)
}
\arguments{
\item{update}{an update function}

\item{comm}{an mpi comm which this function will use for send/receive operations}

\item{swapDelay}{swaps will be attempted every 2*swapDelay+1 iterations \link{deprecated}}
}
\value{
an mcmc closure m(parMCMC,N,eps) that implicitly uses the supplied update function
}
\description{
this version of the MCMC function returns a Markov chain closure
that assumes that it is bein run in an MPI context: R was launched
using \code{runmpi} and the Rmpi package is installed. The chains shall
communicate using the provided \code{comm}.
}
