% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmcInit}
\alias{mcmcInit}
\title{Initialize the Markov chain}
\usage{
mcmcInit(
  parMCMC,
  simulate,
  logLikelihood,
  gradLogLikelihood = NULL,
  fisherInformation = NULL
)
}
\arguments{
\item{parMCMC}{a plain starting value for the Markov chain}

\item{logLikelihood}{a function that maps simulations to
logLikelihood values}

\item{gradLogLikelihood}{the gradient function of the logLikelihood
(optional) -- only if the algorithm requires it}

\item{fisherInformation}{a function that calculates the Fisher Information matrix}
}
\value{
the same starting parameter vector, but with attributes.
}
\description{
This function must append all required attributes to the MCMC
varible, for the Markov chain to update correctly.
}
