% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{model_from_tsv}
\alias{model_from_tsv}
\title{model_from_tsv loads the content from a series of tsv files}
\usage{
model_from_tsv(src = ".")
}
\arguments{
\item{src}{either a vector of files, or a directory with tsv files}
}
\value{
a list of data.frames, one per file, named like the files.
}
\description{
The argument can be either a series of tsv file-names, or a
directory with tsv files. If it is a directory, all tsv files
therein wil be used.
}
