% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{modify<-}
\alias{modify<-}
\title{Modifies a value}
\usage{
modify(x, i = seq_along(NROW(x)), j = seq_along(NCOL(x)), sign = +1) <- value
}
\arguments{
\item{x}{a numeric value to be modified}

\item{i}{row-indices of \code{x} to be modified}

\item{j}{column-indices of \code{x} to be modified}

\item{sign}{modification}
}
\value{
\code{x+v}
}
\description{
This function does the same as \code{x <- x + sign*value}, but without
repeating x modify(x) <- 1 will add one to it. This is a
replacement for the x += 1 syntax of C. This function exists only for
aesthetic reasons.
}
\details{
Specifically, this function should work for matrices.
}
\examples{
 x <- matrix(seq(12),3,4)
 modify(x,seq(2),seq(2)) <- 10
 print(x)
}
