% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{modify<-}
\alias{modify<-}
\title{Modifies a value}
\usage{
modify(x, i = seq(NROW(x)), j = seq(NCOL(x)), sgn = +1) <- value
}
\arguments{
\item{x}{a numeric value to be modified}

\item{i}{row-indices of \code{x} to be modified}

\item{j}{column-indices of \code{x} to be modified}

\item{sgn}{modification}

\item{value}{a numeric value of appropriate size, depending on \code{i} and \code{j}}
}
\value{
The value of x is modified additively ihn place: \code{x <- x + sgn*value}
}
\description{
This function does the same as \code{x <- x + sign*value}, but without
repeating \code{x}. The expression \code{modify(x) <- rnorm(length(x),0,1)}
will add Gaussian noise to it. This is meant as a replacement for
the \verb{x += 1} syntax of C, it exists only for aesthetic reasons.
}
\details{
Specifically, this function should work for matrices, and it is
possible to supply row and column index vectors: \code{x[i,j]} will be
modified.

This function is quite useful if \code{x} has a very long name,
e.g. \code{experiments[[1]]$func}.
}
\examples{
 x <- matrix(seq(12),3,4)
 modify(x,seq(2),seq(2)) <- 10
 print(x)
}
