% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{parameterConversion}
\alias{parameterConversion}
\title{Returns information about parameter conversion}
\usage{
parameterConversion(unit, reactants, products, kinetic.law)
}
\arguments{
\item{unit}{of the kinetic parameters (a character vector), named.}

\item{reactants}{a list of the stoichiometric constants of each
reaction, a list of integer vectors}

\item{products}{a list of the stoichiometric constants of each
reaction, a list of integer vectors}
}
\value{
a vector of parameter conversion factors
}
\description{
Given parameters of a reaction kinetic coefficient model with
concentrations as state variables, this function returns the
parameters of the stochastic version of that model
}
\details{
For this function, it is important that "2 A -> B" is not written
as "A + A -> B" (this may be fixed later).
}
