% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{parameterConversion}
\alias{parameterConversion}
\title{Returns information about parameter conversion}
\usage{
parameterConversion(value, unit, stoichiometry)
}
\arguments{
\item{value}{parameter values (kinetic rate coefficients), a vector}

\item{unit}{of the kinetic parameter, a character vector}

\item{stoichiometry}{a list of the stoichiometric constants of each reaction, a list of integer vectors}

\item{LV}{Avogadro's constant multiplied with the system's volume (the default is a reasonable value for a small volume)}
}
\value{
a vector of parameter conversion factors
}
\description{
Given parameters of a reaction kinetic coefficient model with
concentrations as state variables, this function returns the
parameters of the stochastic version of that model
}
\details{
For this function, it is important that "2 A -> B" is not written
as "A + A -> B" (this may be fixed later).
}
