% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concise.R
\name{parse_concise}
\alias{parse_concise}
\title{Read Concise Error Notation}
\usage{
parse_concise(v, use.errors = requireNamespace("errors"))
}
\arguments{
\item{v}{a character vector of numbers in concise error notation}

\item{use.errors}{if TRUE, the errors package will be used to
retrun an object of type "errors" (from that package).}
}
\value{
a matrix of values and uncertainties (2 rows), or errors object
}
\description{
Convert a vector of strings of the form: c("1.2(3)E-4","1.2(3)E-2") to a matrix with
two rows: values, and uncertainties.
}
\details{
Cincise error notation means that a floating point number is
followed by an integer in parentheses which indicates the
uncertainty of the last digits of the value:
1.2345(12) = 1.2345 ± 0.0012.

If the errors package is installed, then it will be used to
represent the return value.
}
