% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPosterior.R
\name{pcDist}
\alias{pcDist}
\title{plots a sample in parallel coordinates}
\usage{
pcDist(x, prior, color = rgb(0.5, 0.5, 0.5, 0.05), ...)
}
\arguments{
\item{x}{a matrix, with N rows (sample-members), and M columns
(different model parameters). The columns must be named.}

\item{prior}{a data.frame with at least $median, and $stdv
columns. This data.frame may also include the fields: color,
and colorOutline to change the prior error-bars.}

\item{color}{the color of the sample lines, should have some
transparency.}

\item{...}{parameters are passed to matplot.}
}
\value{
produces a plot
}
\description{
This function makes a plot that is quite similar to parallel
coordinates. It includes information about the prior as error-bars,
centered around th eprior's median.
}
