% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationsAndData.R
\name{plot.experiments}
\alias{plot.experiments}
\title{plot function for experiments}
\usage{
\method{plot}{experiments}(x, y, ...)
}
\arguments{
\item{x}{experiment setup (a list)}

\item{y}{simulation results (a list)}

\item{...}{forwarded to the more specific plot function \link[errors:plot.errors]{plot.errors}}
}
\value{
plot object
}
\description{
This function uses plot.errors and the base plot functions like
\link{matplot}.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- write_and_compile(as_ode(m))
ex <- experiments(m,o)
s <- simulator.c(ex,o)
p0 <- values(m$Parameter)
y <- s(p0)
plot(y,ex)
}
