% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCalibration.R
\name{preCalibration}
\alias{preCalibration}
\title{Determine a starting value for ABC's delta}
\usage{
preCalibration(
  objectiveFunction,
  npc = 1000,
  rprior,
  rep = 1,
  p = 0.05,
  sfactor = 0.1,
  delta = 0.01,
  num = 1
)
}
\arguments{
\item{objectiveFunction}{function that, given a (vectorial)
parameter as input, (1) simulates the model with the given parameter, and (2) outputs the
distance between experimental data and simulated data simulated.}

\item{npc}{sample size of pre-calibration.}

\item{rprior}{a function that generates random ABC variables,
distributed according to the prior.}

\item{rep}{number of repetitions of the preCalibration process.}

\item{p}{fraction (top scoring) of sampled points to base Sigma on (Sigma is the covariance matrix for the moves proposed in the ABCMCMC algorithm).}

\item{sfactor}{scales Sigma up or down (Sigma is the covariance matrix for the moves proposed in the ABCMCMC algorithm).}

\item{delta}{ABC threshold.}

\item{num}{number of different starting parameter vectors (initial states of the chains) to generate. Usually, num is equal to the number of chain that will be run in the sampling procedure.}
}
\value{
list with entries prePar (sampled parameters),  preDelta (distances between experimental data and trajectories produced with each of the parameters in prePar), Sigma (covariance matrix for the moves proposed in the ABCMCMC algortihm) and startPar (starting parameters for the ABCMCMC chains)
}
\description{
In ABC settings a model solution is compared to data with an
acceptance threshold: delta. This-pre calibration function attempts
to adjust this delta value.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
cme <- as_cme(m)
ex <- experiments(m)
C <- generate_code(cme)
c_path(cme) <- write_c_code(C)
so_path(cme) <- shlib(cme)
p0 <- log10(values(m$Parameter))
s <- simstoch(ex,cme,parMap=log10ParMap)
Obj <- makeObjective(ex,s)
rprior <- rNormalPrior(p0,0.5)
PC <- preCalibration(Obj,500,rprior,delta=1)
print(names(PC))
print(PC$Sigma)
}
