% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCalibration.R
\name{preCalibration}
\alias{preCalibration}
\title{Determine a starting value for ABC's delta}
\usage{
preCalibration(objectiveFunction, npc = 1000, rprior, rep = 1)
}
\arguments{
\item{objectiveFunction}{function that, given a (vectorial)
parameter as input, simulates the model, and outputs the
distance between experimental data and data simulated from the
model with the parameter provided in input}

\item{npc}{sample size of pre-calibration}

\item{rprior}{a function that generates random ABC variables,
distributed according to the prior}

\item{rep}{number of repetitions of the preCalibration process}
}
\value{
list with entries preDelta and prePar, final values of
calibration run
}
\description{
In ABC settings a model solution is compared to data with an
acceptance threshold: delta. This-pre calibration function attempts
to adjust this delta value.
}
