% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreCalibration.R
\name{preCalibration}
\alias{preCalibration}
\title{Determine a starting value for ABC's delta}
\usage{
preCalibration(
  objectiveFunction,
  npc = 1000,
  rprior,
  nCores = parallel::detectCores(),
  rep = 1
)
}
\arguments{
\item{npc}{sample size of pre-calibration.}

\item{rprior}{a function that generates random ABC variables,
distributed according to the prior.}

\item{nCores}{number of processor cores to use in mclapply().}

\item{rep}{number of repetitions of the preCalibration process}

\item{experiments}{a list of simulation experiments. Same as for
ABCMCMC.}

\item{modelName}{this name will be used to find the file and
functions within the file according to naming conventions (the
model name and file can differ, the file-name can be attached
to the name as a comment).}

\item{parMap}{a remapping function that takes ABC sampling
variables and returns valid model parameters}

\item{getScore}{a function that maps the model's output values to
ABC score values (in comparison to data).}
}
\value{
list with entries preDelta and prePar, final values of
    calibration run
}
\description{
In ABC settings a model solution is compared to data with an
acceptance threshold: delta. This-pre calibration function attempts
to adjust this delta value.
}
