% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{print.cme}
\alias{print.cme}
\title{Print a Summary about the CME model}
\usage{
\method{print}{cme}(x, ...)
}
\arguments{
\item{x}{a model created by \link{as_cme}}

\item{...}{requirement of print generic, not used.}
}
\value{
Nil
}
\description{
This information printed on screen omits the deatls about the
interactions, only the lengths of the vectors included in the data
structure \code{CME}.
}
\examples{
f <- uqsa_example("AKAR4")
m <- model_from_tsv(f)
cmeModel <- as_cme(m)
print(cmeModel)
}
