% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mcmcVariable}
\alias{print.mcmcVariable}
\title{print information about the mcmc variable}
\usage{
\method{print}{mcmcVariable}(x, ...)
}
\arguments{
\item{x}{the variable}

\item{...}{requirement of print generic, not used.}
}
\value{
called for side-effect (printout); no value.
}
\description{
Some mcmc variables have many attributes, which clutter the screen
when accidentally printed. This function prevents these long
printouts.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
c_path(o) <- write_c_code(generate_code(o))
so_path(o) <- shlib(o)
ex <- experiments(m,o)
dprior <- dNormalPrior(values(m$Parameter),m$Parameter$stdv)
s <- simfi(ex,o)
p <- mcmc_init(1.0,values(m$Parameter),s,dprior=dprior)
print(p)
}
