% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{print.simulation}
\alias{print.simulation}
\title{prints the simulation results}
\usage{
\method{print}{simulation}(x, ...)
}
\arguments{
\item{x}{simulation results}

\item{...}{requirement of print generic, not used.}
}
\value{
called for side-effect (printout); no value.
}
\description{
The results, if accidentally printed, are difficult to read.  This
function prevents these accidental prints. It summarizes the
results instead.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
ex <- experiments(m,o)
c_path(o) <- write_c_code(generate_code(o))
so_path(o) <- shlib(o)
s <- simfi(ex,o)
y <- s(values(m$Parameter))
print(y)
}
