% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{print.unit_of_measurement}
\alias{print.unit_of_measurement}
\title{Prints an interpretation string of a unit}
\usage{
\method{print}{unit_of_measurement}(x, ...)
}
\arguments{
\item{x}{an object of type 'unit_of_measurement'}

\item{...}{required by the generic print function.}
}
\value{
called for the side-effect; no value.
}
\description{
The unit object is a tagged data frame, with these columns:
\itemize{
\item multiplier
\item kind
\item scale
\item exponent
}
}
\details{
The interpretation is the same as in SBML units. This function also
prints an inferred unit id: a string that has no special characters
in it and can be used in places where such characters are not
allowed (e.g. SBML unit \code{id} attribute).

The original string that a unit was derived from is attached to the
unit object as a \link{comment}.

Units are produced by the function \link{unit.from.string}.
}
\examples{
lapply(lapply(c("km/h","s^-2","1/s"),unit.from.string),print)
}
