% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{replace_powers}
\alias{replace_powers}
\title{replace_powers}
\usage{
replace_powers(v)
}
\arguments{
\item{v}{a character vector}
}
\description{
replace_powers
}
\examples{
print(replace_powers(c("2^3.1","10^-6","x^2")))
}
