% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{rNormalPrior}
\alias{rNormalPrior}
\title{rNormalPrior returns a random vector generator}
\usage{
rNormalPrior(mean, sd)
}
\arguments{
\item{mean}{mean of the random variables (a vector)}

\item{sd}{standard deviation of the random variables (same size vector as
mean)}
}
\value{
an independentent multivariate normal random vector generating function: rprior(n),
    where n is the requested number of vectors (rows)
}
\description{
The return value is a function that generates random vectors of the
same size as mean and sd from a multivariate normal distribution with 
independent components with mean "mean" and standard deviation "sd".
The random vectors are returned as n
rows of a matrix, where n is the only argument of the returned
function.
}
\examples{
rnp<-rNormalPrior(mean=c(0,1,2),sd=c(1,2,3))
rnp(12)
            [,1]       [,2]       [,3]
[1,]  0.99368106  1.0074638  0.2411801
[2,] -1.39127288  0.3608953 -0.7704979
[3,] -0.23957518 -0.4376285  7.2133888
[4,]  0.55225848 -3.0625581  2.1468860
[5,] -1.74950066 -1.9512875 -0.1923920
[6,]  0.36377999 -0.2222019  4.0832047
[7,]  0.62605893 -0.8562764 -0.2743055
[8,] -0.99829859  2.8863743  7.4753572
[9,] -0.71302121  2.8879101  1.4377714
[10,]  0.04040756 -1.6638517  0.7293233
[11,] -0.04951403  1.9498103  2.1706586
[12,]  0.49326957  2.7241179  7.4083735
}
