% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{rUniformPrior}
\alias{rUniformPrior}
\title{rUniformPrior returns a random vector generator}
\usage{
rUniformPrior(ll, ul)
}
\arguments{
\item{ll}{lower limit of the random variables (a vector)}

\item{ul}{upper limit of the random variables (same size vector as
ll)}
}
\value{
a uniform random vector generating function: runiform(n),
where n is the requested number of vectors (rows)
}
\description{
The return value is a function that generates random vectors of the
same size as ll and ul from a uniform distribution within the
limits defined by ul and ll. The random vectors are returned as n
rows of a matrix, where n is the only argument of the returned
function.
}
\examples{
rup<-rUniformPrior(ll=c(0,1,2),ul=c(1,2,3))
rup(12)
}
