% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity.graph}
\alias{sensitivity.graph}
\title{plot the sensitivity matrix}
\usage{
sensitivity.graph(
  u,
  S,
  color = hcl.colors(dim(S)[2]),
  line.color = hcl.colors(dim(S)[2] + 1),
  do.sort = TRUE,
  decreasing = FALSE,
  ...
)
}
\arguments{
\item{u}{the values of the x-axis for the plot, if named, the names
are put at the tick-marks}

\item{S}{the sensitivity matrix as returned by \code{globalSensitivity()},
S\link{i,j} is with respect to model output i and parameter j}

\item{color}{the list of colors to use for the shaded areas, e.g.:
rainbow(24)}

\item{line.color}{the color of the lines drawn between the shaded
areas}

\item{do.sort}{the parameter sensitivities are sorted according to
the mean over all outputs, the parameter with the most
sensitivity is plotted first, at the bottom}

\item{decreasing}{direction of sort, the first item in the sorted
list (the parameter) will be plotted first, and thus at the
bottom of the plot}

\item{...}{passed on to plot}
}
\value{
nothing
}
\description{
Produce a cumulative shaded area plot for the sensitivity matrix.
}
