% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity.graph}
\alias{sensitivity.graph}
\title{plot the sensitivity matrix}
\usage{
sensitivity.graph(
  u,
  S,
  color = rainbow(dim(S)[2]),
  do.sort = TRUE,
  title = "Sensitivity"
)
}
\arguments{
\item{u}{the values of the x-axis for the plot}

\item{S}{the sensitivity matrix as returned by `sensitivity()`}
}
\value{
nothing
}
\description{
Produce a cumulative shaded area plot for the sensitivity matrix.
}
