% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{shlib}
\alias{shlib}
\title{Compile C code to shared library}
\usage{
shlib(file)
}
\arguments{
\item{file}{the c file that is to be compiled, OR an ode object
with a c.file defined and recorded in it.}
}
\description{
Calls \verb{R CMD SHLIB} to create the model's shared library.
}
