% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{shlib}
\alias{shlib}
\title{Compile C code to shared library}
\usage{
shlib(file)
}
\arguments{
\item{file}{the c file that is to be compiled, OR an ODE/CME object
with a c.file defined and recorded in it.}
}
\value{
the path of the created shared library
}
\description{
Calls \verb{R CMD SHLIB} to create the model's shared library.
}
\details{
The first argument can be a raw character scalar with just the path
of the c code to be compiled, or alternatively an object that has
this information stored within it. The models returned by \link{as_cme}
and \link{as_ode} can both carry this information, attach it via
\link{c.path<-}.
}
