% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPosterior.R
\name{showPosterior}
\alias{showPosterior}
\title{showPosterior makes a pairs plot for a sample}
\usage{
showPosterior(posterior, prior, lower.panel = lp, upper.panel = up, ...)
}
\arguments{
\item{posterior}{a matrix, each row is a sample member}

\item{prior}{a matrix of the same size as the posterior}
}
\description{
This function will display the difference between the posterior and
prior by plotting the posterior as shaded density plots and the
prior as contour lines of level sets. If the two are identical, the
lines will be invisible as they blend into the density
plot. Otherwise the contour lines will show up as a distinct
feature.
}
