% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{simfi}
\alias{simfi}
\title{This creates a closure that simulates the model, similar to simulator.c}
\usage{
simfi(experiments, modelName, parMap = identity, method = 0, omit = 0)
}
\arguments{
\item{experiments}{a list of experiments to simulate: inital
values, inputs, time vectors, initial times}

\item{modelName}{a string (with optional comment indicating an .so
file) which points out the model to simulate}

\item{parMap}{the model will be called with parMap(parABC); so any
parameter transformation can happen there.}

\item{method}{the integration method as an integer (higher numbers
are simpler methods, lower numbers are more advanced methods, 0
maps to 'msbdf')}

\item{omit}{integer, omit optional return values, in this order:
Fisher Information, gradient of the log-likelihood, the
log-likelihood, output functions. Omission includes all
previous entries. \code{omit = 1} omits only the Fisher Information,
\code{omit=3}, omits FI, grad-ll, and log-likelihood calculations.}

\item{parABC}{the parameters for the model, subject to change by
parMap.}
}
\value{
a closure that returns the model's output for a given
parameter vector, and approximate sensitivity matrices, for
each state variable, function, time-point, and parameter
vector.
}
\description{
This is a shorter alternative to simulator.c (C backend). It also
returns the log-likelihood, Fisher Information, and the gradient of
the log-likelihood, under the assumption that the measurement error
is Gaussian. No attempt is made to parallelize this call, all
simulations will be done in sequence.
}
\details{
It returns a closure around:
- experiments,
- the model, and
- parameter mapping

The returned function depends only on parABC (the sampling
parameters). The simulation will be done suing the rgsl backend.

This version of the function does not use the parallel package at
all and cannot add noise to the simulations (unlike simulator.c).
}
\examples{
\donttest{
  f <- uqsa_example("AKAR4")
  m <- model_from_tsv(f)
  o <- as_ode(m)
  ex <- experiments(m,o)
  C <- generateCode(o)
  cat(C,sep='\n',file='./AKAR4_gvf.c')
  modelName <- checkModel(comment(m),'./AKAR4_gvf.c')
  s <- simfi(ex,modelName)
  y <- s(values(m$Parameter)) # simulates
}
}
