% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{simple.unit}
\alias{simple.unit}
\title{Simple unit from string}
\usage{
simple.unit(u = NULL)
}
\arguments{
\item{u}{a unit with no fractions or products}
}
\value{
a data.frame with the unit's properties
}
\description{
This function takes a simple, human readable unit (without '*' or '/'), from a string
and returns a data.frame with the unit's meaning.
}
\details{
In this context, a simple unit is just a prefix, a unit kind, and an exponent, e.g. cm^2
A not-simple unit is: m/s, kg\emph{m/s^2, kg}h
}
