% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{simulator.c}
\alias{simulator.c}
\title{This creates a closure that simulates the model}
\usage{
simulator.c(
  experiments,
  modelName,
  parMap = identity,
  noise = FALSE,
  omit = 3,
  method = 0
)
}
\arguments{
\item{experiments}{a list of experiments to simulate: inital
values, inputs, time vectors, initial times}

\item{modelName}{a string (with optional comment indicating an .so
file) which points out the model to simulate}

\item{parMap}{the model will be called with parMap(parABC); so any
parameter transformation can happen there.}

\item{noise}{boolean variable. If \code{noise=TRUE}, Gaussian noise is
added to the output of the simulations. The standard deviation
of the Gaussian noise is equal to the measurement error. If
\code{noise=FALSE} the output is the deterministic solution of the ODE
system. noise and sensitivity calculations are mutually
exclusive.}

\item{omit}{\code{omit=0} returns all optional return values form the
simulator, \code{omit=1} will not calculate the fisher information
(and thus not return it), \code{omit=2} will omit the gradient of
the log-likelihood, and \code{omit=3} will omit the likelihood
calculations alltogether. Omission is cumulative: \code{omit=3}
omits all the previously mentioned optional quantities.}

\item{parABC}{the parameters for the model, subject to change by
parMap.}
}
\value{
a closure that returns the model's output for a given
parameter vector
}
\description{
This function will use the \link[parallel:mclapply]{parallel::mclapply} to do the
simulations simultaneously. Set \code{options(mc.cores=detectCores())}
or a similar sensible value: \code{options(mc.cores=length(experiments))}
}
\details{
It returns a closure around:
- experiments,
- the model, and
- parameter mapping

The returned function depends only on the parameter vector (or
matrix if more than one simulation per experiment is desired). The
parameter vector this simnulator accepts is probably derived from
the sampling space of a Bayesian method θ, so in the list of
arguments, it is called \code{parABC} or (parMCMC would also have been a
valid choice). These sampling parameters can be mapped to values
the simulator can use via \code{parMap}. \code{parModel <- parMap(parABC)},
where the ODE model is expected to work with \code{parModel}.

Some return values are optional and omiting them saves time.
}
\examples{
 #  model.sbtab <- SBtabVFGEN::sbtab_from_tsv(dir(pattern="[.]tsv$"))
 #  experiments <- SBtabVFGEN::sbtab.data(model.sbtab)
 #  parABC <- SBtabVFGEN::sbtab.quantity(model.sbtab$Parameter)

 #  modelName <- checkModel("<insert_model_name>_gvf.c")
 #  simulate <- simulator.c(experiments, modelName,  parABC)
 #  yf <- simulate(parABC)
}
