% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_StochasticSimulation.R
\name{simulator.stoch}
\alias{simulator.stoch}
\title{Function that creates a closure that simulates a stochastic trajectory with the Gillespie algorithm
given certain experimental conditions and a parameter vector, and computes the
distance between the simulation and the experimental data}
\usage{
simulator.stoch(
  experiments,
  model.tab = model.tab,
  reactions = NULL,
  parMap = identity,
  outputFunction = function(t, state, param) {
     state
 },
  vol = 4e-16,
  unit = 1e-06,
  nStochSim = 3,
  distance = NULL
)
}
\arguments{
\item{experiments}{a list of experiment.}

\item{model.tab}{an SBtab model.}

\item{reactions}{a list of GillespieSSA2 reactions.}

\item{parMap}{a function that translates ABC variables (parABC)
into something the model will accept.}

\item{outputFunction}{a function that, given (t,state,param), outputs the quantity that is measured in the experiments (e.g., the outputFunction may output the concentration of one of the compounds in the system).}

\item{vol}{Volume in which the reactions take place.}

\item{unit}{unit of measure for the volume (it converts the volume into liters).}

\item{nStochSim}{number of stochastic simulations to average over.}

\item{distance}{(optional) function that computes the distance between experimental data and simulated data. Default value is NULL.}
}
\value{
a function that, given a parameter, returns a simulated trajectory obtained via the Gillespie algorithm. Specifically, experimental time points, output value measured at these time points, and (if distance is not NULL) the distance between the simulated trajectories and the experimental data.
}
\description{
Function that creates a closure that simulates a stochastic trajectory with the Gillespie algorithm
given certain experimental conditions and a parameter vector, and computes the
distance between the simulation and the experimental data
}
