% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{smmala_update}
\alias{smmala_update}
\title{SMMALA Update is an MCMC update function}
\usage{
smmala_update(
  simulate,
  logLikelihood = ll,
  dprior = function(x) prod(dnorm(x)),
  gradLogLikelihood = gllf(log10ParMapJac),
  gprior = function(x) (-x),
  fisherInformation = fi(log10ParMapJac),
  fisherInformationPrior = 0,
  parAcceptable = function(p) all(is.finite(p))
)
}
\arguments{
\item{simulate}{a function that simulates the model}

\item{logLikelihood}{a function that returns the log-likelihood
value given the paramegter value, with simulations attached to
the parameter as an attreibute (probably a closure)}

\item{dprior}{a function that returns the prior density of the
given parameter vector}

\item{gradLogLikelihood}{any function that calculates or estimates
the gradient of the log-likelihood function, for the chosen
parameter mapping. Function must take one argument (the MCMC
variable)}

\item{gprior}{a function that returns the gradient of the log-prior
distribution.}

\item{fisherInformation}{a function that estimates the Fisher
Information for a given MCMC variable (parMCMC).}

\item{fisherInformationPrior}{a constant fisherInformation of the
prior distribution (or rather, the precision of the prior)}

\item{parAcceptable}{a function that can be used to reject a
proposal based on the values of the parameters alone (shortcut
to rejection, sans simulation)}
}
\description{
During Markov chain Monte Carlo a given parameter needs to be
updated, the model needs to be simulated at the updated point.
}
\details{
Using the simulations, and an acceptance rule, the proposed update
is either accepted or rejected.

This function returns a closure \code{smmala}, with only \code{parMCMC}
as it's sole argument: \code{parProposal <- smmala(parGiven)}

An optional argument to this function is \code{parAcceptable}, during
sampling, when \code{metropolis} is called as the update function, and
\code{parAcceptable(parProposal)} returns \code{FALSE}, then metropolis
shortcuts to \code{retrun(parGiven)} without performing simulations.

This function can be used to weed out parameter combinations that
would result in obviously nonsensical simulations without wasting
CPU-time.

The argument \code{fisherInformationPrior} is really the precision of
the prior (a constant matrix). It's role is additive to the
fisherInformation and is used to regularize the \emph{final} Fisher
Information Matrix (makes it invertible).
}
