% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{so_path}
\alias{so_path}
\title{Retrieve information about compiled code}
\usage{
so_path(o)
}
\arguments{
\item{o}{the ODE, or CME model}
}
\value{
modified o, with information about compiled code
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- as_ode(m)
c_path(o) <- write_c_code(generate_code(o))
so_path(o) <- shlib(o)
print(so_path(o))
}
\description{
Returns the path of the shared library (.so file).
The model is typically a list of named arrays and matrices.
}
