% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{so_path}
\alias{so_path}
\title{Add information about compiled code}
\usage{
so_path(o)
}
\arguments{
\item{o}{the ODE, or CME model}
}
\value{
modified o, with information about compiled code
}
\description{
Adds the path of the shared library (.so file) to the ODE model.
The model is typically a list of named arrays and matrices.
}
