% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplified_systems_biology_tables.R
\name{standard_error_matrix}
\alias{standard_error_matrix}
\title{Standard Error Matrix from an errors object}
\usage{
standard_error_matrix(M)
}
\arguments{
\item{M}{a matrix with errors (uncertainties)}
}
\value{
A matrix similar to E, with standard error values
}
\description{
If a matrix has an \code{errors} attribute, it is usually a vector.
THis function returns the values of this attribute as a matrix (it
preserves the dimensions of the host matrix).
}
\examples{
M <- matrix(seq(12),3,4,dimnames=liest(letters[seq(3)],LETTERS[seq(4)]))
errors(M) <- abs(M*0.1 + 0.1)
E <- standard_error_matrix(M)
print(E)
}
