% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gillespie.R
\name{stoichiometry}
\alias{stoichiometry}
\title{This function returns a list of named stoichiometric vectors}
\usage{
stoichiometry(formulaList)
}
\arguments{
\item{formulaList}{a list of split reaction formula tokens}
}
\value{
named numeric vector of stoichiometric coefficients
}
\description{
Given an already split list of entries such as c("3 A","B"), this
function returns a numeric vector c(3,1) with names c("A","B").
}
