% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{swap_points}
\alias{swap_points}
\title{Swap the end-points of two Markov chains}
\usage{
swap_points(parMCMC)
}
\arguments{
\item{parMCMC}{a list of Markov chain end points}
}
\value{
a list with some members swapped
}
\description{
This is a conditional swap, according to the rules of parallel
tempering. This function is only useful if the Markov chains have
returned to the global scope and one process will make the decision
and perform the swap.
}
