% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{tune_step_size}
\alias{tune_step_size}
\title{Find a good Step-Size for a given MCMC Algorithm}
\usage{
tune_step_size(
  MCMC,
  parMCMC = attr(MCMC, "init"),
  target_acceptance = 0.25,
  iter.max = 6,
  h = 1e-04
)
}
\arguments{
\item{MCMC}{a Markov chain Monte Carlo closure (function)}

\item{parMCMC}{initial position of the Markov chain, has to be
initialized with \link{mcmc_init}.}

\item{target_acceptance}{a scalar value for the desired acceptance
rate, some algorithms are most efficient with 20\% to 30\%
acceptance, some work well with a very high acceptance.}

\item{iter.max}{maximum number of iterations until the function has
to return.}

\item{h}{initial guess for the MCMC step size}
}
\value{
optimal step size
}
\description{
Given a closure \code{MCMC(p,N,eps)}, where \code{p} is the initial
Markov-chain position, \code{N} a sample-size, and \code{eps} a step-size,
this function finds a good value for eps.
}
\details{
It will take 100 sample points repeatedly, until an acceptance of
\code{target_acceptance} is reached (defaults to 25\%). The step-size is
decreased if acceptance is very low and increased when it is too
high.

This function will do at most
}
\examples{
\donttest{\
m <- model_from_tsv(uqsa_example("AKAP79"))
rwm <- high_level_metropolis(m) # "random walk", metropolis algorithm
p <- rwm \%@\% "init"             # a valid starting point
h <- tune_step_size(rwm,p)
N <- 200
smallSample <- rwm(rwm \%@\% "init",N,h)
print(h)
plot(
  smallSample \%@\% "logLikelihood",
  type="l",
  main=sprintf("step size: \%g",h),
  xlab="iterations",
  ylab="log-likelihood"
)
}
}
