% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqsa-package.R
\docType{package}
\name{uqsa-package}
\alias{uqsa}
\alias{uqsa-package}
\title{uqsa: Uncertainty Quantification and Global Sensitivity Analysis}
\description{
In the field of systems biology, chemical reaction networks are modeled in various ways, two of those are: (i) stochastic simulations (e.g. Gillespie algorithm) and (ii) ordinary differential equations. In this package we use a simple tabular model description of reaction systems and automatically generate C code for either solver type. We use the ordinary differential equation solvers from the GNU Scientific Library and provide an interface that deals with lists of simulation experiments. Each simulation experiment contains both the data, and instructions for the model to replicate the data. We use approximate Bayesian computation methods (combined with Markov chain Monte Carlo and sequential Monte Carlo, particle filters) as well as classic methods such as Random Walk Metropolis (Gaussian transition kernel) and Simplified Manifold Metropolis adjusted Langevin algorithm for a Bayesian investigation of the model´s parameter space. Experiments can be evaluated in a sequence; intermediate probability densities are modeled using the 'VineCopula' package. The package is also intended to be useful in an HPC environment, with some functions that use 'pbdMPI' capabilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://icpm-kth.github.io/uqsa/}
}

}
\author{
\strong{Maintainer}: Andrei Kramer \email{andreikr@kth.se}

Authors:
\itemize{
  \item Alexandra Jauhiainen \email{Alexandra.Jauhiainen@astrazeneca.com}
  \item Olivia Eriksson \email{olivia@kth.se} [contributor, copyright holder]
  \item Federica Milinanni \email{fedmil@kth.se}
}

}
\keyword{internal}
