% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqsa-package.R
\docType{package}
\name{uqsa-package}
\alias{uqsa}
\alias{uqsa-package}
\title{uqsa: Uncertainty Quanitification and Global Sensitivity Analysis via ABC sampling}
\description{
This package will solve an initial value problem (given an ordinary differential equation) for a list of simulation experiments. The simulations will be compared to data, repeatedly. The package can sample the parameters of the model via ABC (approximate Bayesian computation) and by doing so quantify the uncertainty that remains within the model even after considering the data. Experiments are evaluated in a sequence, intermediate probability densities are modeled using the VineCopua package.
}
\author{
\strong{Maintainer}: Olivia Eriksson \email{olivia@mech.kth.se}

Authors:
\itemize{
  \item Alexandra Jauhiainen \email{Alexandra.Jauhiainen@astrazeneca.com}
  \item Federica Milinanni \email{fedmil@kth.se}
  \item Andrei Kramer \email{andreikr@kth.se} [contributor]
}

}
\keyword{internal}
