% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uqsa_example_data.R
\name{uqsa_example}
\alias{uqsa_example}
\title{Load an example model for this package}
\usage{
uqsa_example(
  modelName = NULL,
  full.names = TRUE,
  pattern = "[.]tsv$",
  f = NULL
)
}
\arguments{
\item{modelName}{name of model, e.g.: "AKAR4", "AKAP79", "CaMKII";
if empty, this function lists all available examples.}

\item{full.names}{return full paths to files - defaults to TRUE}

\item{pattern}{pattern to find specific files; if \code{NULL}, this function
returns the directory of the example}

\item{f}{file ending, search for file endings in \code{f}, alternative to \code{pattern}}
}
\value{
The location of the examples in the current environment if
called with no arguments, the paths to the model files if a
modelName was provided or the full path to the example if the
file pattern \emph{pattern} is unset
}
\description{
This function finds the path to an example model, given by name.
In the SBtab format, model and data travel together (in different
tables, but the same documents).
}
\details{
By default this function returns the names of the tsv files
belonging to the named model. If no modelName is provided it
returns possible names (contents of the top-level example
directory).
}
\examples{
uqsa_example()
uqsa_example("AKAR4",full.names=FALSE)
uqsa_example("AKAP79",f='R',full.names=FALSE)
uqsa_example("AKAP79",pat="^run.*R$")
}
