% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_and_compile}
\alias{write_and_compile}
\title{Writes code to file and compiles}
\usage{
write_and_compile(o)
}
\arguments{
\item{o}{ode model for which code is generated and written to a file}
}
\value{
a copy of \code{o} with file paths added to it
}
\description{
This function accepts an ode model, or cme model, generates code,
compiles it to a shared library, and returns a changed object.
possibly changed by the user. It writes the contents to a c file
named 'modelName_gvf.c'. This file is compiled to './modelName.so'
using normal command line tools, not \verb{R CMD SHLIB}
}
\details{
This entire function can be replaced with a call to \code{cat()} and
then compiling the written file in the system's shell.
}
\examples{
m <- model_from_tsv(uqsa_example("AKAR4"))
o <- write_and_compile(as_ode(m))
print(o)
}
