% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{write_c_code}
\alias{write_c_code}
\title{Write the C code to a file}
\usage{
write_c_code(
  C,
  model.name = comment(C),
  file = file.path(tempdir(), digest::digest(C, "xxh3_64"), paste0(model.name, ".c"))
)
}
\arguments{
\item{C}{the code to write, as a character array.}

\item{model.name}{a string with no special characters, will be used in the file name}

\item{file}{override the default file name (based on hashing)}
}
\value{
the path of the written file
}
\description{
This function does not compile the code, it only writes it to a
file in a temporary location (tempdir). By default, the name of the
file will contain the hash of the entire code.
}
