% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{yacasMath}
\alias{yacasMath}
\title{yacasMath converts math to Ryacas compatible math}
\usage{
yacasMath(v, reverse = FALSE)
}
\arguments{
\item{v}{a chcarcter vector with math expressions}

\item{reverse}{do the reverse operation}
}
\value{
a character vector with yacas math
}
\description{
Given a string like "exp(2\emph{x)" this function returns a string that
yacas can process: "Exp(2}x)"
}
